### 2022-12-25 #568 @maRce10
ohun
Submitting Author Name: Marcelo Araya-Salas
Submitting Author Github Handle: @maRce10
Repository:  https://github.com/maRce10/ohun
Version submitted: 0.1.0
Submission type: Standard

Version accepted: TBD
---

```
Package: ohun
Title: Optimizing Acoustic Signal Detection
Version: 0.1.0
Maintainer: Marcelo Araya-Salas <marcelo.araya@ucr.ac.cr>
Description: Facilitates the automatic detection of acoustic signals,
  providing functions to diagnose and optimize the performance of detection
  routines. Detections from other software can also be explored and optimized.
  Reference: Hossin & Sulaiman (2015) <doi:10.5121/ijdkp.2015.5201>.
VignetteBuilder: knitr, rmarkdown
Repository: CRAN
Authors@R: c(person("Marcelo", "Araya-Salas", role = c("aut", "cre"), email = "marcelo.araya@ucr.ac.cr", comment = c(ORCID = "0000-0003-3594-619X")))
Date/Publication: 2016-04-19 08:12:11
Imports: pbapply, viridis, crayon, methods, stats, utils, seewave (>= 2.0.1), RCurl, fftw, knitr, rjson, rlang, sp, igraph, Sim.DiffProc
Depends: R (>= 3.2.1), tuneR, warbleR (>= 1.1.28)
Suggests: rmarkdown, testthat (>= 3.0.0), formatR

```

The packages allows to get data from animal acoustic signal recordings in an automated manner

Scientific community working with bioacoustic data. Useful to automate analysis of animal acoustic signals for a variety of research questions (e.g. behavior, ecology, evolution, conservation)

- the [warbleR](https://marce10.github.io/warbleR/) package has similar functions for detecting but the implementation is not as friendly (I am also the maintainer) and has no tools for detection diagnose/optimization. [monitoR](https://cran.r-project.org/web/packages/monitoR/index.html) also implements one of the detection approaches in ohun but no diagnosing/optimization tools are supplied.
