+++ tests/testthat/test-internal_functions.R
@@ -1,0 +1,14 @@
+test_that("stop2", {
+
+
+  expect_error(stop2("stop!"))
+
+})
+
+
+test_that("stop2", {
+
+
+  expect_warning(warning2("warning!"))
+
+})
