+++ R/ohun-package.R
@@ -20,8 +20,6 @@
 #' All functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency. The package works on sound files in '.wav', '.mp3', '.flac' and '.wac' format.
 #'
 #' @import warbleR
-#' @import rjson
-#' @import RCurl
 #' @import pbapply
 #' @import tuneR
 #' @import seewave
