+++ NAMESPACE
@@ -18,7 +18,6 @@ export(split_acoustic_data)
 export(summarize_diagnostic)
 export(template_correlator)
 export(template_detector)
-import(RCurl)
 import(Sim.DiffProc)
 import(fftw)
 import(grDevices)
@@ -26,7 +25,6 @@ import(graphics)
 import(methods)
 import(parallel)
 import(pbapply)
-import(rjson)
 import(seewave)
 import(stats)
 import(tuneR)
