+++ codecov.yml
@@ -1,13 +1,8 @@
 comment: false
-
 language: R
-
 sudo: false
-
 cache: packages
-
 after_success:
-
 - Rscript -e 'covr::codecov()'
 
 coverage:
