### 2022-12-30 c9c7bf9
added info pkgcheck.yaml
