+++ pkgcheck.yaml
@@ -18,11 +18,11 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {os: macOS-latest,   r: 'release'}
-          - {os: windows-latest, r: 'release'}
-          - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
+          # - {os: macOS-latest,   r: 'release'}
+          # - {os: windows-latest, r: 'release'}
+          # - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest,   r: 'release'}
-          - {os: ubuntu-latest,   r: 'oldrel-1'}
+          # - {os: ubuntu-latest,   r: 'oldrel-1'}
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
@@ -30,40 +30,40 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
 
     steps:
-      - uses: actions/checkout@v2
-
-      - uses: r-lib/actions/setup-pandoc@v2
-
-      - uses: r-lib/actions/setup-r@v2
-        with:
-          r-version: ${{ matrix.config.r }}
-          http-user-agent: ${{ matrix.config.http-user-agent }}
-          use-public-rspm: true
+      # - uses: actions/checkout@v2
+      #
+      # - uses: r-lib/actions/setup-pandoc@v2
+      #
+      # - uses: r-lib/actions/setup-r@v2
+      #   with:
+      #     r-version: ${{ matrix.config.r }}
+      #     http-user-agent: ${{ matrix.config.http-user-agent }}
+      #     use-public-rspm: true
 
       - name: Extra deps
         if: runner.os == 'Linux'
         run: |
           sudo apt-get install -y make
+#
+#       - name: ctags latest
+#         if: runner.os == 'Linux'
+#         run: |
+#           git clone https://github.com/universal-ctags/ctags.git
+#           cd ctags
+#           ./autogen.sh
+#           ./configure --prefix=/usr
+#           make
+#           sudo make install
 
-      - name: ctags latest
-        if: runner.os == 'Linux'
-        run: |
-          git clone https://github.com/universal-ctags/ctags.git
-          cd ctags
-          ./autogen.sh
-          ./configure --prefix=/usr
-          make
-          sudo make install
-
-      - uses: r-lib/actions/setup-r-dependencies@v2
-        with:
-           extra-packages: |
-             any::rcmdcheck
-
-      - run: rappdirs::user_cache_dir()
-        shell: Rscript {0}
-
-      - uses: r-lib/actions/check-r-package@v2
-        with:
-          upload-snapshots: true
-
+      # - uses: r-lib/actions/setup-r-dependencies@v2
+      #   with:
+      #      extra-packages: |
+      #        any::rcmdcheck
+      #
+      # - run: rappdirs::user_cache_dir()
+      #   shell: Rscript {0}
+      #
+      # - uses: r-lib/actions/check-r-package@v2
+      #   with:
+      #     upload-snapshots: true
+      #
