### 2022-12-31 c2788f4
update pkgcheck.yaml
