+++ pkgcheck.yaml
@@ -43,7 +43,7 @@ jobs:
       - name: Extra deps
         if: runner.os == 'Linux'
         run: |
-          sudo apt-get install -y make fftw3
+          sudo apt-get install -y make libssl-dev libcurl4-openssl-dev libxml2-dev libfontconfig1-dev libharfbuzz-dev libfribidi-dev libfreetype6-dev libpng-dev libtiff5-dev libjpeg-de cmake libfftw3-3 libfftw3-dev libsndfile1 libsndfile1-dev
           sudo apt-add-repository -r ppa:armagetronad-dev/ppa
           sudo apt update -q
 #
