+++ .circleci/config.yml
@@ -7,6 +7,7 @@ jobs:
     environment:
     docker:
       - image: rocker/verse
+    parallelism: 4
     steps:
       - checkout
 
