+++ pkgdown.yml
@@ -1,0 +1,10 @@
+pandoc: 2.19.2
+pkgdown: 2.0.7
+pkgdown_sha: ~
+articles:
+  ohun: ohun.html
+last_built: 2022-12-31T22:30Z
+urls:
+  reference: https://marce10.github.io/ohun/reference
+  article: https://marce10.github.io/ohun/articles
+
