+++ README.md
@@ -28,18 +28,18 @@ optimize detection routines.
 
 The main features of the package are:
 
--   The use of reference annotations for detection diagnostic and
-    optimization
--   The use of signal detection theory indices to evaluate detection
-    performance
+- The use of reference annotations for detection diagnostic and
+  optimization
+- The use of signal detection theory indices to evaluate detection
+  performance
 
 The package offers functions for:
 
--   Curate references and acoustic data sets
--   Diagnose detection performance
--   Optimize detection routines based on reference annotations
--   Energy-based detection
--   Template-based detection
+- Curate references and acoustic data sets
+- Diagnose detection performance
+- Optimize detection routines based on reference annotations
+- Energy-based detection
+- Template-based detection
 
 All functions allow the parallelization of tasks, which distributes the
 tasks among several processors to improve computational efficiency. The
