### 2022-12-31 58669fc
adding macOS to circleCI
