+++ .github/workflows/tic.yml
@@ -30,9 +30,9 @@ jobs:
         config:
           # use a different tic template type if you do not want to build on all listed platforms
           - { os: windows-latest, r: "release" }
-          - { os: macOS-latest, r: "release", pkgdown: "true", latex: "true" }
-          - { os: ubuntu-latest, r: "devel" }
-          - { os: ubuntu-latest, r: "release" }
+          # - { os: macOS-latest, r: "release", pkgdown: "true", latex: "true" }
+          # - { os: ubuntu-latest, r: "devel" }
+          # - { os: ubuntu-latest, r: "release" }
 
     env:
       # make sure to run `tic::use_ghactions_deploy()` to set up deployment
@@ -57,6 +57,14 @@ jobs:
         if: matrix.config.latex == 'true'
 
       - uses: r-lib/actions/setup-pandoc@v2
+
+
+      - name: Extra deps
+        if: runner.os == 'Linux'
+        run: |
+          sudo apt-get install -y make libssl-dev libcurl4-openssl-dev libxml2-dev libfontconfig1-dev libharfbuzz-dev libfribidi-dev libfreetype6-dev libpng-dev libtiff5-dev libjpeg-dev cmake libfftw3-3 libfftw3-dev libsndfile1 libsndfile1-dev flac
+          sudo apt-add-repository -r ppa:armagetronad-dev/ppa
+          sudo apt update -q
 
       # set date/week for use in cache creation
       # https://github.community/t5/GitHub-Actions/How-to-set-and-access-a-Workflow-variable/m-p/42970
