+++ .github/workflows/tic.yml
@@ -29,7 +29,7 @@ jobs:
       matrix:
         config:
           # use a different tic template type if you do not want to build on all listed platforms
-          # - { os: windows-latest, r: "release" }
+          - { os: windows-latest, r: "release" }
           - { os: macOS-latest, r: "release", pkgdown: "true", latex: "true" }
           # - { os: ubuntu-latest, r: "devel" }
           - { os: ubuntu-latest, r: "release" }
