+++ pkgcheck.yaml
@@ -1,70 +1,0 @@
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
-
-name: R-CMD-check
-
-jobs:
-  R-CMD-check:
-    runs-on: ${{ matrix.config.os }}
-
-    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
-
-    strategy:
-      fail-fast: false
-      matrix:
-        config:
-          - {os: macOS-latest,   r: 'release'}
-          - {os: windows-latest, r: 'release'}
-          # - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
-          # - {os: ubuntu-latest,   r: 'release'}
-          # - {os: ubuntu-latest,   r: 'oldrel-1'}
-
-    env:
-      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-      R_KEEP_PKG_SOURCE: yes
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-
-    steps:
-      # - uses: actions/checkout@v2
-      #
-      # - uses: r-lib/actions/setup-pandoc@v2
-      #
-      # - uses: r-lib/actions/setup-r@v2
-      #   with:
-      #     r-version: ${{ matrix.config.r }}
-      #     http-user-agent: ${{ matrix.config.http-user-agent }}
-      #     use-public-rspm: true
-
-      - name: Extra deps
-        if: runner.os == 'Linux'
-        run: |
-          sudo apt-get install -y make libssl-dev libcurl4-openssl-dev libxml2-dev libfontconfig1-dev libharfbuzz-dev libfribidi-dev libfreetype6-dev libpng-dev libtiff5-dev libjpeg-dev cmake libfftw3-3 libfftw3-dev libsndfile1 libsndfile1-dev flac
-          sudo apt-add-repository -r ppa:armagetronad-dev/ppa
-          sudo apt update -q
-#
-#       - name: ctags latest
-#         if: runner.os == 'Linux'
-#         run: |
-#           git clone https://github.com/universal-ctags/ctags.git
-#           cd ctags
-#           ./autogen.sh
-#           ./configure --prefix=/usr
-#           make
-#           sudo make install
-
-      # - uses: r-lib/actions/setup-r-dependencies@v2
-      #   with:
-      #      extra-packages: |
-      #        any::rcmdcheck
-      #
-      # - run: rappdirs::user_cache_dir()
-      #   shell: Rscript {0}
-      #
-      # - uses: r-lib/actions/check-r-package@v2
-      #   with:
-      #     upload-snapshots: true
