+++ .github/workflows/tic.yml
@@ -31,7 +31,7 @@ jobs:
           # use a different tic template type if you do not want to build on all listed platforms
           - { os: windows-latest, r: "release" }
           # - { os: macOS-latest, r: "release", pkgdown: "true", latex: "true" }
-          # - { os: ubuntu-latest, r: "devel" }
+          - { os: ubuntu-latest, r: "devel" }
           # - { os: ubuntu-latest, r: "release" }
 
     env:
@@ -81,7 +81,7 @@ jobs:
           key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
           restore-keys: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
 
-    - name: "non R windows dependencies"
+      - name: "non R windows dependencies"
         if: runner.os == 'Windows'
         run: |
           Set-ExecutionPolicy Bypass -Scope Process -Force; [System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072; iex ((New-Object System.Net.WebClient).DownloadString('https://community.chocolatey.org/install.ps1'))
