### 2023-01-01 8b9e9dc
remove template detector test
