### 2023-01-01 b452b3b
cleaning ci (only ubuntu and macOS)
