+++ .github/workflows/test-coverage.yaml
@@ -19,7 +19,20 @@ jobs:
 
       - uses: r-lib/actions/setup-r@v2
         with:
-          use-public-rspm: true
+          r-version: ${{ matrix.config.r }}
+          Ncpus: 4
+
+      - uses: r-lib/actions/setup-tinytex@v2
+        if: matrix.config.latex == 'true'
+
+      - uses: r-lib/actions/setup-pandoc@v2
+
+      - name: Extra deps
+        if: runner.os == 'Linux'
+        run: |
+          sudo apt-get install -y make libssl-dev libcurl4-openssl-dev libxml2-dev libfontconfig1-dev libharfbuzz-dev libfribidi-dev libfreetype6-dev libpng-dev libtiff5-dev libjpeg-dev cmake libfftw3-3 libfftw3-dev libsndfile1 libsndfile1-dev flac
+          sudo apt-add-repository -r ppa:armagetronad-dev/ppa
+          sudo apt update -q
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
