+++ .github/workflows/test-coverage.yaml
@@ -21,6 +21,7 @@ jobs:
         with:
           r-version: ${{ matrix.config.r }}
           Ncpus: 4
+          use-public-rspm: true
 
       - uses: r-lib/actions/setup-tinytex@v2
         if: matrix.config.latex == 'true'
