### 2023-01-12 #568 @maRce10
Hi

thanks for reviewing my package. Not sure why you got problems with coverage. It shows 78% coverage in codecov:

https://app.codecov.io/github/maRce10/ohun

as well as in the coverage status badge

https://github.com/maRce10/ohun

So I am not sure how to proceed now.

Cheers

Marcelo
