### 2023-01-12 c1e108d
replace with seq_len and vapply when posible, remove unused dependencies/suggests
