### 2023-01-12 #568 @maRce10
I replaced with seq_len() and vapply() when possible, replace cat() with message() and removed unused dependencies/suggests: [6ea380d](https://github.com/maRce10/ohun/commit/6ea380d706dfeb705684245ca2460b552edfa4ae)
