### 2023-01-12 #568 @mpadge
@maRce10 It took a bit of digging, but the test failures are because `flac` is not installed. I see you're doing explicit installs in your own GitHub action file. It's common to put these kind of "non-standard" system dependencies in the "SystemRequirements" field of your DESCRIPTION file. Our system would then pick that up and install it, but without that has no way of knowing which libraries your package might need. That said, most systems only match entries given there to the ["rules" directory of rstudio/r-system-requirements](https://github.com/rstudio/r-system-requirements). Any libraries not listed there, including "flac", are simply ignored. What you should at least do is:

1. List all libraries which do appear there, and which your package requires, in your DESCRIPTION file.
2. Let us know of any libraries on which your package depends are which are not listed there , and we'll add them to our build system. (Or you couuld submit a PR to [this Dockerfile](https://github.com/ropensci-review-tools/pkgcheck/blob/main/Dockerfile).)

Many of the libraries given in your actions file will automatically pull in several of the others, so you should reduce this set to the minimal possible length, either through checking disribution listings (at least for Debian or Ubuntu), or iterating in a docker container.

Please ping here when your DESCRIPTION has been updated, and let us know what else we might need to install. Thanks.
