### 2023-01-13 #568 @maRce10
OK thanks. These non-standard dependencies are not from the ohun package but from a package it relies heavily on, warbleR. These dependencies are now included in the DESCRIPTION file of warbleR on github. Should  I also add them to ohun?
