### 2023-01-16 8e07ae4
remove fftw3 sox
