### 2023-01-17 f7bcba0
Update test-template_detector.R
