### 2023-01-18 62cdc87
update test energy_detector
