+++ tests/testthat/test-template_detector.R
@@ -1,6 +1,6 @@
 test_that("1 template", {
 
-  if (Sys.info()[1] != "Windows"){
+
   data(lbh1, package = "ohun")
   data(lbh2, package = "ohun")
   tuneR::writeWave(lbh1, file.path(tempdir(),  "lbh1.wav"), extensible = FALSE) #save sound files
@@ -21,14 +21,11 @@ test_that("1 template", {
   expect_true(is_selection_table(td))
 
   expect_true(nrow(td) == 22)
-  }
-  expect_true(TRUE)
 
 })
 
 
 test_that("2 templates", {
-  if (Sys.info()[1] != "Windows"){
   data(lbh1, package = "ohun")
   data(lbh2, package = "ohun")
   tuneR::writeWave(lbh1, file.path(tempdir(),  "lbh1.wav"), extensible = FALSE) #save sound files
@@ -57,6 +54,5 @@ test_that("2 templates", {
 
   expect_true(is_selection_table(td))
 
-  expect_true(nrow(td) == 42)} else
-    expect_true(TRUE)
+  expect_true(nrow(td) == 42)
 })
