### 2023-01-19 #568 @jhollist
@maRce10 just now getting a chance to take a look at this.

One question I have relates to the submission checklist.  You did not check the "I have read the [author guide](https://devdevguide.netlify.app/authors-guide.html) and I expect to maintain this package for at least 2 years or to find a replacement."  Was curious if that was an oversight or if you expect maintenance of this package to be short term only.
