### 2023-01-20 #568 @jhollist
I think we are ready to keep moving on this.  Thanks for updating the checklist, @maRce10.

Couple of things I would like to see addressed prior to review are:

- Replace use of `sp` with `sf`.  The `sp` package, while not going away entirely, is being replaced by `sf`.  I took a very quick look at your code and it looks like there are only a few spots where you are using `sp`.
- Start adressing some of the `goodpractice` results in https://github.com/ropensci/software-review/issues/568#issuecomment-1386966949.  In particular the `lintr` suggestions are some low hanging fruit (don't get too worked up on the over 80 character lines).   All of these don't need to be addressed, but considering making these changes is a good idea.

As you work on these I will start to identify some possible reviewers.  Any questions, just let me know.
