+++ R/internal_functions.R
@@ -5,7 +5,7 @@ stop2 <- function (...)
   stop(..., call. = FALSE)
 }
 
-
+# warning function that doesn't print call
 warning2 <- function (...)
 {
   warning(..., call. = FALSE)
@@ -17,30 +17,201 @@ message2 <- function(x, color = "black")
   message(colortext(x, as = color))
 
 
-colortext <- function(text, as = c("red", "blue", "green", "magenta", "cyan", "orange", "black", "silver")) {
-  if (has_color()) {
-    unclass(cli::make_ansi_style(ohun_style(as))(text))
-  } else {
-    text
+colortext <-
+  function(text,
+           as = c("red",
+                       "blue",
+                       "green",
+                       "magenta",
+                       "cyan",
+                       "orange",
+                       "black",
+                       "silver")) {
+    if (has_color()) {
+      unclass(cli::make_ansi_style(ohun_style(as))(text))
+    } else {
+      text
+    }
   }
-}
 
+# to check if string includes color
 has_color <- function() {
   cli::num_ansi_colors() > 1
 }
 
-ohun_style <- function(color = c("red", "blue", "green", "magenta", "cyan", "orange", "black", "silver")) {
-  type <- match.arg(color)
+# to print text in colors
+ohun_style <-
+  function(color = c("red",
+                          "blue",
+                          "green",
+                          "magenta",
+                          "cyan",
+                          "orange",
+                          "black",
+                          "silver")) {
+    type <- match.arg(color)
 
-  c(
-    red = "red",
-    blue = "blue",
-    green = "green",
-    magenta = "magenta",
-    cyan = "cyan",
-    orange = "orange",
-    black = "black",
-    silver = "silver"
-  )[[color]]
-}
+    c(
+      red = "red",
+      blue = "blue",
+      green = "green",
+      magenta = "magenta",
+      cyan = "cyan",
+      orange = "orange",
+      black = "black",
+      silver = "silver"
+    )[[color]]
+  }
 
+
+
+# internal function to find (and plot) templates in the acoustic space
+find_templates <-
+  function(reference = NULL,
+           n.sub.spaces,
+           space = NULL,
+           plot = TRUE,
+           color = "#21908C4D",
+           xlab = "Dimension 1",
+           ylab = "Dimension 2")
+  {
+    x <- rep(1, n.sub.spaces)
+    space <- space[, 1:2]
+    space <- data.frame(space)
+    space$...NROW <- seq_len(nrow(space))
+
+    # mean center
+    mean_dim1 <- mean(space[, 1])
+    mean_dim2 <- mean(space[, 2])
+
+    # center at 0
+    space[, 1] <-  space[, 1] - mean_dim1
+    space[, 2] <-  space[, 2] - mean_dim2
+
+    x <- c(0, cumsum(x) / sum(x))
+    dx <- diff(x)
+    nx <- length(dx)
+    radius <- 100
+
+    t2xy <- function(t) {
+      t2p <- 2 * pi * t + 0 * pi / 180
+      list(x = radius * cos(t2p), y = radius * sin(t2p))
+    }
+
+    if (plot)
+      plot(
+        space[, 1] + mean_dim1,
+        space[, 2] + mean_dim2,
+        pch = 20,
+        cex = 2,
+        col = color,
+        xlab = xlab,
+        ylab = ylab
+      )
+
+    # get polygon for each sub space
+    if (n.sub.spaces > 1)
+      polys <- lapply(1:nx, function(i) {
+        n <- max(2, floor(100 * dx[i])) # 100 equivalent to edges in pie()
+        P <- t2xy(seq.int(x[i], x[i + 1], length.out = n))
+
+        if (plot)
[…]
