+++ R/get_templates.R
@@ -107,108 +107,3 @@ c("Dimension 1", "Dimension 2")
 
     return(templates)
   }
-
-
-# internal function to find (and plot) templates in the acoustic space
-find_templates <- function(reference = NULL, n.sub.spaces, space = NULL, plot = TRUE, color = "#21908C4D", xlab = "Dimension 1", ylab = "Dimension 2")
-{
-  x <- rep(1, n.sub.spaces)
-  space <- space[, 1:2]
-  space <- data.frame(space)
-  space$...NROW <- seq_len(nrow(space))
-
-  # mean center
-  mean_dim1 <- mean(space[, 1])
-  mean_dim2 <- mean(space[, 2])
-
-  # center at 0
-  space[, 1] <-  space[, 1] - mean_dim1
-  space[, 2] <-  space[, 2] - mean_dim2
-
-  x <- c(0, cumsum(x)/sum(x))
-  dx <- diff(x)
-  nx <- length(dx)
-  radius <- 100
-
-  t2xy <- function(t) {
-    t2p <- 2 * pi * t + 0 * pi/180
-    list(x = radius * cos(t2p), y = radius * sin(t2p))
-  }
-
-  if (plot)
-  plot(space[, 1] + mean_dim1, space[, 2] + mean_dim2, pch = 20, cex = 2, col = color, xlab = xlab, ylab = ylab)
-
-  # get polygon for each sub space
-  if (n.sub.spaces > 1)
-  polys <- lapply(1:nx, function(i){
-    n <- max(2, floor(100 * dx[i])) # 100 equivalent to edges in pie()
-    P <- t2xy(seq.int(x[i], x[i + 1], length.out = n))
-
-    if (plot)
-    lines(c(P$x[1], 0) + mean_dim1, c(P$y[1], 0) + mean_dim2, lty = 3, col = "gray", lwd = 3)
-
-    pol_df <- data.frame(x = c(P$x, 0), y = c(P$y, 0))
-    poly <- sp::Polygons(list(sp::Polygon(pol_df)), ID = 1)
-
-    spp <- sp::SpatialPolygons(list(poly))
-
-    return(spp)
-  }) else polys <- vector()
-
-  # get  centroids, the first one is the centroid of the enitre space
-  centroids <- vector(length = length(polys) + 1)
-
-  for (i in 0:length(polys)){
-
-    if (i > 0)
-      whch_within  <- which(!is.na(sp::over(x = sp::SpatialPoints(space[, 1:2]),  y = sp::SpatialPolygons(polys[[i]]@polygons), returnList = FALSE))) else whch_within <- seq_len(nrow(space))
-
-      if (length(whch_within) > 0){
-        if (length(whch_within) > 1) {
-
-          sub_space <- space[whch_within, ]
-
-          # get centroid
-          centroid_coors <- colMeans(sub_space[, 1:2])
-
-          # and distance to centroid
-          dists_to_centroid <-
-            unlist(lapply(seq_len(nrow(sub_space)), function(x)
-              stats::dist(
-                rbind(sub_space[x, 1:2], centroid_coors)
-              )))
-
-          centroid <- sub_space$...NROW[which.min(dists_to_centroid)]
-        } else
-          centroid <- sub_space$...NROW[whch_within]
-      } else
-        centroid <- NA
-
-      centroids[i + 1] <- centroid
-  }
-
-
-  if (length(centroids) > 1)
-    if (any(centroids[-1] == centroids[1])){
-      message2(color = "silver", x = paste(sum(centroids[-1] == centroids[1]),
-                         "sub-space centroid(s) coincide with the overall centroid and was (were) removed"
-      ))
[…]
