+++ NAMESPACE
@@ -39,6 +39,9 @@ importFrom(igraph,graph_from_incidence_matrix)
 importFrom(igraph,max_bipartite_match)
 importFrom(methods,is)
 importFrom(rlang,call_args)
+importFrom(sf,st_as_sf)
+importFrom(sf,st_intersects)
+importFrom(sf,st_polygon)
 importFrom(stats,cor)
 importFrom(stats,weighted.mean)
 importFrom(tuneR,writeWave)
