### 2023-01-23 #568 @maRce10
Thanks!

I replaced sp with sf so sp is not longer used. I also dealt with most `goodpractice` results. After the first check round I fixed all possible instances in which `seq_len()` was a better option and replaced `sapply()` by `vapply()` when possible.

`bioacoustics` also has 1 method for doing automatic detection, but has no tools for diagnosing and optimizing those detections. So `ohun` could actually be used as a complimentary tool  for `bioacoustics`. `bioacoustics` also does other things that `ohun` doesn't, like feature extraction and acoustic file manipulation
