+++ NEWS.md
@@ -1,0 +1,19 @@
+<!-- based on: https://devguide.ropensci.org/newstemplate.html#newstemplate -->
+
+ohun 0.1.1 (developmental)
+=========================
+
+### NEW FEATURES
+
+### MINOR IMPROVEMENTS
+
+  * `sp` package replaced by `sf`
+  * Replace `sapply()` with `vapply()`
+
+
+ohun 0.1.0 (2022-12-19)
+=========================
+
+### NEW FEATURES
+
+  * released to CRAN
