### 2023-01-24 #568 @maRce10
OK, I added NEWS.md and the rOpenSci status badge (although doesn't seem to show up properly)
