### 2023-02-20 #568 @sammlapp
I'm working on my review, but would like to check in on `devtools::check('ohun')`. I've never run devtools in R, so forgive me if I'm doing something wrong here. Running `devtools::test('ohun')` in R Studio completes without errors, showing all tests passed. However, running `devtools::check('ohun')` in R Studio results in an error. I have a MackBook Pro  running Mac OS 12.6 with R Studio 2022.07.1 Build 554 and the following `version` info

```               _
platform       aarch64-apple-darwin20
arch           aarch64
os             darwin20
system         aarch64, darwin20
status
major          4
minor          2.1
year           2022
month          06
day            23
svn rev        82513
language       R
version.string R version 4.2.1 (2022-06-23)
nickname       Funny-Looking Kid
```

Full traceback of `devtools::check('ohun')` :
```

Documenting ════════════════════════════════════════════════════════════════════════════════════════
ℹ Updating ohun documentation
ℹ Loading ohun

══ Building ═══════════════════════════════════════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/Users/SML161/ohun/DESCRIPTION’ (351ms)
─  preparing ‘ohun’:
E  creating vignettes (16s)
   --- re-building ‘ohun.Rmd’ using rmarkdown
   Loading required package: tuneR
   Loading required package: warbleR
   Loading required package: seewave
   Loading required package: NatureSounds
   Loading required package: knitr
   Quitting from lines 326-353 (ohun.Rmd)
   Error: processing vignette 'ohun.Rmd' failed with diagnostics:
   must install 'Sim.DiffProc' to use this function
   --- failed re-building ‘ohun.Rmd’

   SUMMARY: processing the following file failed:
     ‘ohun.Rmd’

   Error: Vignette re-building failed.
   Execution halted
Error in `(function (command = NULL, args = character(), error_on_status = TRUE, …`:
! System command 'R' failed
---
Exit status: 1
stdout & stderr: <printed>
---
Type .Last.error to see the more details.

.Last.error #outputs

Backtrace:
 1. devtools::check("ohun")
 2. withr::with_envvar(pkgbuild::compiler_flags(FALSE), action = "prefix", …
 3. base::force(code)
 4. pkgbuild::build(pkg$path, tempdir(), args = build_args, quiet = quiet, …
 5. withr::with_temp_libpaths(rcmd_build_tools(options$cmd, c(options$path, …
 6. base::force(code)
 7. pkgbuild::rcmd_build_tools(options$cmd, c(options$path, options$args), …
 8. pkgbuild::with_build_tools({ …
 9. base::withCallingHandlers(callr::rcmd_safe(..., env = env, spinner = FALSE, …
10. callr::rcmd_safe(..., env = env, spinner = FALSE, show = FALSE, …
11. callr:::run_r(options)
12. base::with(options, with_envvar(env, do.call(processx::run, c(list(bin, …
13. base::with.default(options, with_envvar(env, do.call(processx::run, …
14. base::eval(substitute(expr), data, enclos = parent.frame())
15. base::eval(substitute(expr), data, enclos = parent.frame())
16. callr:::with_envvar(env, do.call(processx::run, c(list(bin, args = real_cmdargs, …
17. base::force(code)
18. base::do.call(processx::run, c(list(bin, args = real_cmdargs, stdout_line_callback = real_callback(stdo…
19. (function (command = NULL, args = character(), error_on_status = TRUE, …
20. base::throw(new_process_error(res, call = sys.call(), echo = echo, …
21. | base::signalCondition(cond)
22. (function (e) …
23. asNamespace("callr")$err$throw(e)
```
