### 2023-02-20 #568 @sammlapp
- [ ] **Vignette(s):** demonstrating major functionality that runs successfully locally - I wasn't able to see or access the vignette in R studio with `vignette()`

- [ ] **Performance:** Any performance claims of the software been confirmed. (I did not test parallel computing)

---

### General notes

I don't typically use R, so some things are not obvious to me. Where can I find the full API documentation? I opened the /docs/index.html to view the docs by rendering the HTML locally, but presumably they are online somewhere. I would recommend linking to the documentation near the top of the README.

The use of the word "reference" to mean "audio annotations" or "labels" is unfamiliar to me, I'm not sure if this word is typically used in the related warbleR package.

The package should give the user some expectation for under what circumstances the detection methods are likely to be effective. Although it mentions that high SNR is required for amplitude thresholding and stereotyped sounds are necessary for template matching, there are other requirements of the data for these approaches to be effective.

In this review, I have not delved into the creation of warbleR selection tables and the interaction of the package with such tables. It seems that this package is aimed towards users who already create and use warbleR selection tables, and because that package is established I consider it "legacy" in the sense that I don't want to review or critique it here. In my experiments with the package, I created a dataframe of "selections" using the `split_acoustic_data` function on dry mode (`only.sels=TRUE`), renamed columns, and added random values for annotation bounds just to have a fake set of annotations on local audio files to play with.

Overall, functions generally worked as described in the documentation. However, in my experiments the `get_templates` function gave an error that I coudn't figure out how to avoid. The primary recommendations of this review are

(1) to improve documentation and function/variable/argument names to better match the action that is actually taken by a function;

(2) to consider a moderate refactor of code to increase modularity, reduce redundancy, and aim to follow the Single Responsibility Principle;

(3) increase code readability by breaking complex logical lines into separate lines with additional comments, plus comments describing the general purpose of large blocks of code.

`devtools::test('ohun')`runs without error, but as posted above `devtools::check('ohun')` raises an error for me.

#### Redundancy of functionality:
Template-based cross correlation is implemented in [monitoR](https://cran.r-project.org/web/packages/monitoR/vignettes/monitoR_QuickStart.pdf), while energy-based detection is implemented in [bioacoustics](https://cran.r-project.org/web/packages/bioacoustics/vignettes/introduction.html) (the implementation seems more generic and featured than the one in ohun).

I wonder if the proposed value of Ohun is mostly in its specific ability to interact with the formats of annotations produced by warbleR. If so, the top-level documentation and README could clarify this purpose.

Specifically, when I started using the program without having any experience with warbleR, it was not clear to me how to create annotation tables or generally what the utility of most of the package functions are. Writing a description of the characteristics of data that this package expects as input would be helpful. For instance, "This package provides utilities for comparing detection and annotations of audio events described by frequency and time boxes. Creating annotations containing such frequency-time boxes [is possible in the warbleR package?]."

### Notes on the package design

The package is very 'flat' in the sense that functions rarely call other functions, but instead take a bunch of arguments to do a bunch of things in one go. One result of this is that many functions take a long list of parameters, and parameters are redundant across many functions. For example, several functions take the same set of spectrogram parameters. To me, this raises the concern of accidental inconsistency in a user's script (eg, by passing different spectrogram parameters to two different functions).

In the language of rOpenSci software recommendations, this is a sign that the functions are not "modular". In other places it might be seen as a violation of the Single Responsibility Principle (SRP). For example, the `label_specto` function violates SRP because it doesn't just label a spectrogram: it computes a spectrogram, plots it, then overlays labels, and might additionally comput and plot an amplitude signal. As another example, I get_envelopes.R and energy_detector.R seem to both implement the process of getting an amplitude envelope, rather than using shared modular functionality specific to this purpose.

Though it would take a significant refactoring effort, I believe creating a more modular code base that aims to achieve the SRP could improve the package and user experience overall.

### Notes on the source code

- In general, the code is difficult to read because of the formatting and smushing lots of logic into a single functional "line" (which is broken into many lines anyways due to formatting). Separating out logical steps into sequential lines, with comments where appropriate, would make the code easier to read/understand/maintain. However, it would take a major refactoring effort.

- Consider adding units to variable names eg hold_time_ms

- A few specific comments on two source files provide examples of what I see as possible areas for improvement:

#### split_acoustic_data.R:

- documentation should list supported sound file types
- documentation should list required columns of table
- documentation should say what will happen with incomplete audio clip at end of file
- comments denoting the purpose of larger sections of code are helpful, eg # INPUT VALIDATION
or # Create the audio clips

#### summarize_diagnostic.R:

- The code style makes it difficult to review the logic or infer the operation performed by much of this code.  For example, lines 131-141 of `summarize_diagnostic.R` :

```r
          summ_diagnostic$mean.duration.true.positives <-
            if (any(!is.na(Y$mean.duration.true.positives)))
              round(
                stats::weighted.mean(
                  x = Y$mean.duration.true.positives,
                  w = Y$true.positives,
                  na.rm = TRUE
                ),
                0
              ) else
            NA
```

In general I'm not reading this code base to check for logical errors, but I think that the code style will make it difficult for others (or the authors) to debug, extend, understand or modify the code. In this example, separating the logical operation into a few separate and sequential lines of code could make it easier to digest, and would provide opportunities to add comments for any individual steps whose purposes are not obvious. (For example in this code, a comment could explain why we would want a weighted mean of true positive durations.)

- Input validation is redundant across many functions, eg validating the `cores` argument. Rather than repeating code in several modules, use modular helper functions. That way, if you change something it is consistent across the codebase.

- Comments should explain not just what's being done but also why (rather than re-stating the code in English). For example, instead of `# set bottom and top freq if not supplied` you could explain `the bottom and top frequency are needed to make a spectrogram. if they're not supplied, we set them to 0 and Ns=sample_rate/2`. Another example, in `template_detector.R` L177 "#  if no detections" is not a helpful comment - the comment should explain what will happen and why.

- Since I'm not an R programmer, I might not understand conventions for writing R modules. However, I don't think it's a good idea to define your helper functions inside another function. For example L218 of template_correlator.R begins a definition of a function to create a spectrogram. Even if this is not part of the API (its a hidden or internal function), I think it should be defined outside of the template_correlator function (1) because it doesn't make sense that it would be re-defined each time template_correlator is called, and (2) because it makes reading the code for template_correlator unnecessarily long.

- I'm not sure how namespaces work in R but it would be helpful for a code reader to know where functions are coming from. For instance, `get_templates.R` calls `find_templates` which is not explicitly imported or defined anywhere in the file. It apparently can be accessed because it exists in another .R file in the package, or because all functions from the package have been imported into the default namespace.

### Notes on the Getting Started tutorial page
The "main features" description is jargon-y and could be written in broader language. For example, "The use of signal detection theory indices to evaluate detection performance" -> "Evaluate detection accuracy with precision, recall, and other metrics". As is, I feel like it obscures the functionality of the package. As another example, I wasn't sure what "Optimize detection routines based on reference annotations" meant - after reading further, I think it basically means "Adjust detection parameters to optimize accuracy on a labeled set of recordings"

"Template-based detection" It's not true that template-based detection "doesn’t depend on the signal-to-noise ratio" - high background noise or low signal level will decrease the accuracy of this approach

outputs (eg, tables) in the Getting Started tutorial are very hard to interpret in the current form, could they be rendered as human-readable tables?

"Detections from other software can also be explored and optimized." I don't see how detections from another platform could be "optimized" in ohun unless you just mean changing a score threshold and measuring the resulting precision and recall.

An example of how to analyze a large set of files, for instance all files in a folder-subfolder structure, would be helpful. I had to do some googling to figure out how to do "globbing" ie find all paths mathing a pattern.

### Notes on API documentation and variable/function names

Throughout the codebase, I recommend more informative variable and argument names: for example, in the function `merge_overlaps`, to a new user, pb is not interpreted as "progress bar"

Throughout the documentation: why are arguments that are numbers/strings called "numeric vectors of length 1" and "character vectors of length 1"? This makes me thing I should pass c(0) or [0] instead of 0, for instance, as an argument.

`get_envelopes` :
- smooth argument - should state the specific function/implementation of smoothing used
- paths and files arguments: using the current path as default is fragile and does not match the conventions used elsewhere in the package. I recommend making the argument(s) mandatory. It is also unclear that specifying a folder but not files means that the function should attempt to analyze any/all audio files in the folder (only a specific file extension? how does it know which files to analyze or not?). Is that what happens if files=NULL?

`energy_detector`: again, I don't think making files optional and defaulting to all audio in the working directory is a good idea.

`label_spectro`: can you provide an example of how to use this function with an audio file? If I want to run it on anything other than a simulated array of numbers or the two built-in audio files, I need some prior line of code to load the audio file, for example:
```
library(tuneR)
w <- readWave("~/sample_audio/birds_10s.wav")
```

`bp` argument (in several functions): how is the signal bandpassed? Should state specifically, for example, a 9th order Butterworth bandpass filter

`feature_reference`: "Extract quantitative features of references" - from the function name and short description, I expected this to extract acoustic features of the annotated audio events themselves, rather than summary statistics of the annotations per se. Perhaps the description could be "Summarize temporal and frequency dimensions of annotations and gaps" or something similar. Ideally, the function's name would reflect its functionality, something like `summarize_annotation_boxes`.

- gap times: I tried using this summary to choose a hold.time argument for `energy_detector`, but the min, max, and mean are not very informative, because what matters is the typical gap time between close-together annotations. Maybe the median or 25th percentile, or some other statistic, would be more informative.

Similarly, `feature_acoustic_data` sounds like it would measure some acoustic feature. Since the function returns information about the file types, I would call it something like `summarize_audio_file_characteristics`

`filter_detection` I'm confused as to what this function is doing. The documentation says it "removes ambiguous detections (split and merged detections)", but the detailed documentation should explain what is meant by split and merged detections. I'm also confused whether this is meant to be applied to human-created annotations, automatically-generated detections, or can be used for either purpose.

`optimize_template_detector`: this function compares performance across parameter values, it does not itself optimize the algorithm, so its name is a bit misleading. Perhaps `template_parameter_sweep` is closer. I'm not sure what paramter values it sweeps over after reading the documentation, since it requires the user to have already run `template_correlator`. Also, I think the part of the documentation that says "Those parameters can be later used for performing a more efficient detection using optimize_template_detector." must be a typo, perhaps it means "using `template_correlator`" instead?

`label_spectro()`: this is a plotting function, so it would be better named something like `plot_spec_and_labels`

`get_templates`: can you explain what features are being used to create this principle component analysis of template similarity? Also, after understanding what this function does, it seems like it would be better named something like `select_representative_templates` or `select_representative_annotations` since it chooses representative templates from a set of annotations.

`split_acoustic_data`: it should be possible, if not required, to list the audio files to split. It seems like I must split every file within a folder in the current implementation.

`label_detection` function name does not match what it does. It seems this function _evaluates_ detections against a set of annotations/labels/"references" so perhaps it should be called `evaluate_detections` or `compare_annotations_and_detections`

`template_correlator` it's unclear to me how to use the files argument. I don't know what 'X' is as referred to in the documentation, and I would expect to simply be able to pass a vector of file paths.

`full_spectrograms`: it was surprising to me that this function automatically saved image files to the current working directory, rather than asking for a save dir or displaying them in the console. This doesn't seem like a good practice, because I might not want to save the results and if I do, I'd like to choose where they go.

Other note:
I would be very surprised if template matching is faster than amplitude-based detection, as claimed in the "Improving detection speed" section. Amplitude-based detection is about the lightest task you can perform on a time signal while 2d cross correlation is relatively heavy.

### Errors during usage

When I try to run get_templates on my own file set, I get this error:
```
Error in prcomp.default(spectral_parameters[, 2:27], scale. = TRUE) :
  cannot rescale a constant/zero column to unit variance
```
I tried to run get_templates on a set of local files, but it seems somehow one of the features has no variance and this is causing the PCA to fail. I wasn't able to resolve this.

Running template_correlator then template_detector on the results produced a table containing NA values for many rows in the start, end, and scores columns. I did not expect this, and it caused errors when I tried to run other analyses on these detections.

Once I ran template_correlator and template_detector, I wasn't sure how to inspect my detections although it seemed like I should be able to use full_spectrograms or some other package function.
- I eventually realized that if I used na.omit(selections) to drop rows with NA values, I could run full_spectrograms to look at the results. I wonder if there is a way to more selectively choose to look at some of the detections alone, rather than having to view the entire audio files (imagine if I had a 1 hour audio file with 3 1-second detections).

When I first ran label_spectro, I got the message "Error in loadNamespace(x) : there is no package called ‘viridis’". I'm not sure if this suggests a missing dependency. I resolved the issue by running `install.packages("viridis")`

When I run `label_spectro` with env = TRUE, I get an error message:
```
w <- readWave("~/sample_audio/birds_10s.wav")

label_spectro(wave = w@left,
              f=w@samp.rate,
              env = TRUE)

Error in label_spectro(wave = w@left, f = w@samp.rate, env = TRUE, fastdisp = TRUE) :
  trying to get slot "samp.rate" from an object of a basic class ("integer") with no slots
```

Add a more helpful error message for accidental use of Hz instead of kHz: If my detection table has Hz instead of kHz and I run get_templates, the error message says "Caution! This is a null spectrum. The spectral entropy is null!"
