### 2023-04-21 cc24443
fix bug mulitiple file detection and vignette typos
