+++ R/optimize_template_detector.R
@@ -78,6 +78,7 @@ optimize_template_detector <-
            pb = TRUE,
            by.sound.file = FALSE,
            previous.output = NULL) {
+
     if (!is(template.correlations, "template_correlations"))
       stop2(
         "'template.correlations' must be an object of class 'template_correlations' generated by template_detector()"
@@ -190,6 +191,7 @@ optimize_template_detector <-
 
             diagnostic <- do.call(rbind, sub_diagnostics_list)
 
+            diagnostic <- data.frame(threshold = x, diagnostic)
             return(diagnostic)
           }
         )
@@ -198,11 +200,11 @@ optimize_template_detector <-
       diagnostics <- do.call(rbind, diagnostics_list)
 
       # summarize sound files
-      threshold <-
-        rep(threshold, each = nrow(diagnostics_list[[1]]))
+      # threshold <-
+        # rep(threshold, each = nrow(diagnostics_list[[1]]))
 
       # add thresholds to output
-      diagnostics <- data.frame(threshold = threshold, diagnostics)
+      # diagnostics <- data.frame(threshold = threshold, diagnostics)
 
       # add previous output
       if (!is.null(previous.output))
