### 2023-05-05 20e9921
rm irrelevant references to contributing.md
