### 2023-05-05 b690142
fix alt text ohun not baRulho
