+++ DESCRIPTION
@@ -13,7 +13,6 @@ URL: https://marce10.github.io/ohun/, https://github.com/maRce10/ohun
 BugReports: https://github.com/maRce10/ohun/issues/
 VignetteBuilder: knitr, rmarkdown
 RoxygenNote: 7.2.3
-Repository: CRAN
 Language: en-US
 Authors@R: c(person("Marcelo", "Araya-Salas", role = c("aut", "cre"), email = "marcelo.araya@ucr.ac.cr", comment = c(ORCID = "0000-0003-3594-619X")))
 Date/Publication: 2016-04-19 08:12:11
