### 2023-05-07 f8f0dec
rm date/publication field
