+++ DESCRIPTION
@@ -15,7 +15,6 @@ VignetteBuilder: knitr, rmarkdown
 RoxygenNote: 7.2.3
 Language: en-US
 Authors@R: c(person("Marcelo", "Araya-Salas", role = c("aut", "cre"), email = "marcelo.araya@ucr.ac.cr", comment = c(ORCID = "0000-0003-3594-619X")))
-Date/Publication: 2016-04-19 08:12:11
 Imports: cli, methods, stats, utils, seewave (>= 2.0.1), fftw, rlang, sf, igraph
 Depends: R (>= 3.2.1), tuneR, warbleR (>= 1.1.28)
 Suggests: knitr, rmarkdown, testthat, viridis, Sim.DiffProc
