+++ DESCRIPTION
@@ -15,7 +15,7 @@ VignetteBuilder: knitr, rmarkdown
 RoxygenNote: 7.2.3
 Language: en-US
 Authors@R: c(person("Marcelo", "Araya-Salas", role = c("aut", "cre"), email = "marcelo.araya@ucr.ac.cr", comment = c(ORCID = "0000-0003-3594-619X")))
-Imports: cli, methods, stats, utils, seewave (>= 2.0.1), fftw, rlang, sf, igraph
-Depends: R (>= 3.2.1), tuneR, warbleR (>= 1.1.28)
+Imports: tuneR, warbleR (>= 1.1.28), cli, methods, stats, utils, seewave (>= 2.0.1), fftw, rlang, sf, igraph
+Depends: R (>= 3.2.1)
 Suggests: knitr, rmarkdown, testthat, viridis, Sim.DiffProc
 Config/testthat/edition: 3
