### 2023-05-07 95530fe
rm config/testthat/edition field
