+++ DESCRIPTION
@@ -1,5 +1,4 @@
 Package: ohun
-Type: Package
 Title: Optimizing Acoustic Signal Detection
 Version: 0.1.1
 Maintainer: Marcelo Araya-Salas <marcelo.araya@ucr.ac.cr>
