### 2023-05-09 8698c0b
added CRAN installation
