+++ R/optimize_energy_detector.R
@@ -166,147 +166,147 @@ optimize_energy_detector <-
     #check path to working directory
     if (is.null(path))
       path <- getwd() else
-      if (!dir.exists(path))
-        stop2("'path' provided does not exist") else
-      path <- normalizePath(path)
+        if (!dir.exists(path))
+          stop2("'path' provided does not exist") else
+            path <- normalizePath(path)
 
-    # if files not supplied then used those from reference
-    if (is.null(files))
-      files <- unique(reference$sound.files)
+          # if files not supplied then used those from reference
+          if (is.null(files))
+            files <- unique(reference$sound.files)
 
-    # if 'files' are found in reference
-    if (!any(files %in% reference$sound.files))
-      stop2("Not a single sound file in the working directory is found in 'reference'")
+          # if 'files' are found in reference
+          if (!any(files %in% reference$sound.files))
+            stop2("Not a single sound file in the working directory is found in 'reference'")
 
-    # get all possible combinations of parameters
-    exp_grd <-
-      expand.grid(
-        threshold = threshold,
-        peak.amplitude = peak.amplitude,
-        smooth = smooth,
-        hold.time = hold.time,
-        min.duration = if (is.null(min.duration))
-          - Inf else
-          min.duration,
-        max.duration = if (is.null(max.duration))
-          Inf else
-          max.duration,
-        thinning = thinning
-      )
+          # get all possible combinations of parameters
+          exp_grd <-
+            expand.grid(
+              threshold = threshold,
+              peak.amplitude = peak.amplitude,
+              smooth = smooth,
+              hold.time = hold.time,
+              min.duration = if (is.null(min.duration))
+                - Inf else
+                  min.duration,
+              max.duration = if (is.null(max.duration))
+                Inf else
+                  max.duration,
+              thinning = thinning
+            )
 
-    # if previous output included
-    if (!is.null(previous.output)) {
-      # create composed variable to find overlapping runs
-      previous.output$temp.label <-
-        apply(previous.output[, c(
-          "threshold",
-          "peak.amplitude",
-          "smooth",
-          "hold.time",
-          "min.duration",
-          "max.duration",
-          "thinning"
-        )], 1, paste, collapse = "-")
+          # if previous output included
+          if (!is.null(previous.output)) {
+            # create composed variable to find overlapping runs
+            previous.output$temp.label <-
+              apply(previous.output[, c(
+                "threshold",
+                "peak.amplitude",
+                "smooth",
+                "hold.time",
+                "min.duration",
+                "max.duration",
+                "thinning"
+              )], 1, paste, collapse = "-")
 
-      exp_grd <-
-        exp_grd[!apply(exp_grd[, c(
-          "threshold",
-          "peak.amplitude",
-          "smooth",
-          "hold.time",
-          "min.duration",
-          "max.duration",
-          "thinning"
-        )], 1, paste, collapse = "-") %in% previous.output$temp.label,]
+            exp_grd <-
+              exp_grd[!apply(exp_grd[, c(
+                "threshold",
+                "peak.amplitude",
+                "smooth",
+                "hold.time",
+                "min.duration",
+                "max.duration",
+                "thinning"
+              )], 1, paste, collapse = "-") %in% previous.output$temp.label,]
 
-      # remove composed variable
-      previous.output$temp.label <- NULL
-    }
+            # remove composed variable
+            previous.output$temp.label <- NULL
+          }
 
 
-    if (nrow(exp_grd) == 0) {
-      cat(
-        "all combinations were already evaluated on previous call to this function (based on 'pevious.output')"
-      )
+          if (nrow(exp_grd) == 0) {
+            cat(
[…]
