### 2023-05-23 01b8299
shorten some diagnose output columns
