### 2023-05-23 283e7ba
adde min.overlap and macro.averaging to optimizing functions
