+++ R/optimize_template_detector.R
@@ -2,7 +2,8 @@
 #'
 #' @description \code{\link{optimize_template_detector}} optimizes acoustic template detection
 #' @usage optimize_template_detector(template.correlations, reference, threshold,
-#' cores = 1, pb = TRUE, by.sound.file = FALSE, previous.output = NULL)
+#' cores = 1, pb = TRUE, by.sound.file = FALSE, previous.output = NULL,
+#' macro.average = FALSE, min.overlap = 0.5)
 #' @param template.correlations An object of class 'template_correlations' (generated by \code{\link{template_correlator}}) in which to optimize detections. Must contain data for all sound files as in 'reference'. It can also contain data for additional sound files. In this case the routine assumes that no sound events are found in those files, so detection from those files are all false positives.
 #' @param reference Data frame or 'selection.table' (following the warbleR package format) with the reference selections (start and end of the sound events) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end". \strong{It must contain the reference selections that will be used for detection optimization}.
 #' @param threshold Numeric vector of length > 1 with values between 0 and 1 specifying the correlation threshold for detecting sound event occurrences (i.e. correlation peaks). Must be supplied. \strong{Several values should be supplied for optimization}.
@@ -11,6 +12,8 @@
 #' @param pb Logical argument to control progress bar and messages. Default is \code{TRUE}.
 #' @param by.sound.file Logical to control if diagnostics are calculated for each sound file independently (\code{TRUE}) or for all sound files combined (\code{FALSE}, default).
 #' @param previous.output Data frame with the output of a previous run of this function. This will be used to include previous results in the new output and avoid recalculating detection performance for parameter combinations previously evaluated.
+#' @param macro.average Logical argument to control if diagnostics are first calculated for each sound file and then averaged across sound files, which can minimize the effect of unbalanced sample sizes between sound files. If \code{FALSE} (default) diagnostics are based on aggregated statistics irrespective of sound files. The following indices can be estimated by macro-averaging: overlap, mean.duration.true.positives, mean.duration.false.positives, mean.duration.false.positives, mean.duration.false.negatives, proportional.duration.true.positives, recall and precision (f1.score is always derived from recall and precision). Note that when applying macro-averaging, recall and precision are not derived from the true positive, false positive and false negative values returned by the function.
+#' @param min.overlap Numeric. Controls the minimum amount of overlap required for a detection and a reference sound for it to be counted as true positive. Default is 0.5. Overlap is measured as intersection over union.
 #' @return A data frame in which each row shows the result of a detection job for each cutoff value, including the following diagnostic metrics:
 #' \itemize{
 #'  \item \code{true.positives}: number of sound events in 'reference' that correspond to any detection. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'.
@@ -77,7 +80,9 @@ optimize_template_detector <-
            cores = 1,
            pb = TRUE,
            by.sound.file = FALSE,
-           previous.output = NULL) {
+           previous.output = NULL,
+           macro.average = FALSE,
+           min.overlap = 0.5) {
 
     if (!is(template.correlations, "template_correlations"))
       stop2(
@@ -169,7 +174,9 @@ optimize_template_detector <-
                     detection = detection[detection$template == r,],
                     by.sound.file = by.sound.file,
                     cores = cores,
-                    pb = FALSE
+                    pb = FALSE,
+                    macro.average = macro.average,
+                    min.overlap = min.overlap
                   )
 
                 # get column names
