+++ R/optimize_energy_detector.R
@@ -4,7 +4,8 @@
 #' @usage optimize_energy_detector(reference, files = NULL, threshold = 5,
 #' peak.amplitude = 0, hop.size = 11.6, wl = NULL, smooth = 5, hold.time = 0,
 #' min.duration = NULL, max.duration = NULL, thinning = 1, cores = 1, pb = TRUE,
-#'  by.sound.file = FALSE, bp = NULL, path = ".", previous.output = NULL, envelopes = NULL)
+#'  by.sound.file = FALSE, bp = NULL, path = ".", previous.output = NULL, envelopes = NULL,
+#'  macro.average = FALSE, min.overlap = 0.5)
 #' @param reference Selection table (using the warbleR package's format, see \code{\link[warbleR]{selection_table}}) or data frame with columns
 #' for sound file name (sound.files), selection number (selec), and start and end time of sound event
 #' (start and end). \strong{It must contain the reference selections that will be used for detection optimization}.
@@ -37,6 +38,8 @@
 #'The current working directory is used as default.
 #' @param previous.output Data frame with the output of a previous run of this function. This will be used to include previous results in the new output and avoid recalculating detection performance for parameter combinations previously evaluated.
 #' @param envelopes An object of class 'envelopes' (generated by \code{\link{get_envelopes}}) containing the amplitude envelopes of the sound files to be analyzed. If 'files' and 'envelopes' are not supplied then the function will work on all supported format sound files in the working directory.
+#' @param macro.average Logical argument to control if diagnostics are first calculated for each sound file and then averaged across sound files, which can minimize the effect of unbalanced sample sizes between sound files. If \code{FALSE} (default) diagnostics are based on aggregated statistics irrespective of sound files. The following indices can be estimated by macro-averaging: overlap, mean.duration.true.positives, mean.duration.false.positives, mean.duration.false.positives, mean.duration.false.negatives, proportional.duration.true.positives, recall and precision (f1.score is always derived from recall and precision). Note that when applying macro-averaging, recall and precision are not derived from the true positive, false positive and false negative values returned by the function.
+#' @param min.overlap Numeric. Controls the minimum amount of overlap required for a detection and a reference sound for it to be counted as true positive. Default is 0.5. Overlap is measured as intersection over union.
 #' @return A data frame in which each row shows the result of a detection job with a particular combination of tuning parameters (including in the data frame). It also includes the following diagnostic metrics:
 #' \itemize{
 #'  \item \code{true.positives}: number of sound events in 'reference' that correspond to any detection. Matching is defined as some degree of overlap in time. In a perfect detection routine it should be equal to the number of rows in 'reference'.
@@ -123,7 +126,9 @@ optimize_energy_detector <-
            bp = NULL,
            path = ".",
            previous.output = NULL,
-           envelopes = NULL) {
+           envelopes = NULL,
+           macro.average = FALSE,
+           min.overlap = 0.5) {
     # hopsize
     if (!is.numeric(hop.size) |
         hop.size < 0)
@@ -284,7 +289,9 @@ optimize_energy_detector <-
                     time.diagnostics = TRUE,
                     pb = FALSE,
                     cores = cores,
-                    path = path
+                    path = path,
+                    macro.average = macro.average,
+                    min.overlap = min.overlap
                   )
                 ))
 
