+++ R/split_acoustic_data.R
@@ -23,23 +23,23 @@
 #' @details This function aims to reduce the size of sound files in order to simplify some processes that are limited by sound file size (big files can be manipulated, e.g. \code{\link{energy_detector}}).
 #' @examples
 #' {
-#' #load data and save to temporary working directory
-#' data("lbh1", "lbh2")
-#' writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
-#' writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
+#'   # load data and save to temporary working directory
+#'   data("lbh1", "lbh2")
+#'   writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
+#'   writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
 #'
-#' #split files in 1 s files
-#' split_acoustic_data(sgmt.dur = 1, path = tempdir())
+#'   # split files in 1 s files
+#'   split_acoustic_data(sgmt.dur = 1, path = tempdir())
 #'
-#' # Check this folder
-#' tempdir()
+#'   # Check this folder
+#'   tempdir()
 #' }
 #'
 #' @references {
 #' Araya-Salas, M., Smith-Vidaurre, G., Chaverri, G., Brenes, J. C., Chirino, F., Elizondo-Calvo, J., & Rico-Guevara, A. 2022. ohun: an R package for diagnosing and optimizing automatic sound event detection. BioRxiv, 2022.12.13.520253. https://doi.org/10.1101/2022.12.13.520253
 #' }
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr})
-#last modification on aug-23-2021 (MAS)
+# last modification on aug-23-2021 (MAS)
 split_acoustic_data <-
   function(path = ".",
            sgmt.dur = 10,
@@ -49,78 +49,99 @@ split_acoustic_data <-
            pb = TRUE,
            only.sels = FALSE,
            X = NULL) {
-    #check path to working directory
-    if (is.null(path))
-      path <- getwd() else
-      if (!dir.exists(path))
-        stop2("'path' provided does not exist") else
+    # check path to working directory
+    if (is.null(path)) {
+      path <- getwd()
+    } else if (!dir.exists(path)) {
+      stop2("'path' provided does not exist")
+    } else {
       path <- normalizePath(path)
+    }
 
-    #stop if files is not a character vector
+    # stop if files is not a character vector
     if (!is.null(files) &
-        !is.character(files))
+      !is.character(files)) {
       stop2("'files' must be a character vector")
+    }
 
-    if (is.null(files))
+    if (is.null(files)) {
       files <-
-        list.files(path = path,
-                   pattern = "\\.wav$|\\.wac$|\\.mp3$|\\.flac$",
-                   ignore.case = TRUE) #list .wav files in working director
+        list.files(
+          path = path,
+          pattern = "\\.wav$|\\.wac$|\\.mp3$|\\.flac$",
+          ignore.case = TRUE
+        )
+    } # list .wav files in working director
 
-    #stop if no wav files are found
-    if (length(files) == 0)
+    # stop if no wav files are found
+    if (length(files) == 0) {
       stop2("no sound files in working directory")
+    }
 
     if (!is.null(X)) {
-      #if X is not a data frame
-      if (!any(is.data.frame(X), warbleR::is_selection_table(X)))
+      # if X is not a data frame
+      if (!any(is.data.frame(X), warbleR::is_selection_table(X))) {
         stop2("X is not of a class 'data.frame' or 'selection_table'")
+      }
 
-      if (warbleR::is_extended_selection_table(X))
+      if (warbleR::is_extended_selection_table(X)) {
         stop2("This function cannot take extended selection tables ('X' argument)")
+      }
 
[…]
