+++ R/feature_reference.R
@@ -13,16 +13,16 @@
 #' @name feature_reference
 #' @details The function extracts quantitative features from reference tables that can inform the range of values to be used in a energy-based detection optimization routine. Features related to selection duration can be used to set the 'max.duration' and 'min.duration' values, frequency related features can inform bandpass values, gap related features inform hold time values and duty cycle can be used to evaluate performance.
 #' @examples {
-#' # load data and save example files into temporary working directory
-#' data("lbh1", "lbh2", "lbh_reference")
-#' writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
-#' writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
+#'   # load data and save example files into temporary working directory
+#'   data("lbh1", "lbh2", "lbh_reference")
+#'   writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
+#'   writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
 #'
-#' # summary across sound files
-#' feature_reference(reference = lbh_reference, path = tempdir())
+#'   # summary across sound files
+#'   feature_reference(reference = lbh_reference, path = tempdir())
 #'
-#' # summary across sound files
-#' feature_reference(reference = lbh_reference, by.sound.file = TRUE, path = tempdir())
+#'   # summary across sound files
+#'   feature_reference(reference = lbh_reference, by.sound.file = TRUE, path = tempdir())
 #' }
 #' @seealso \code{\link{optimize_energy_detector}}, \code{\link{optimize_template_detector}}
 #' @author Marcelo Araya-Salas \email{marcelo.araya@@ucr.ac.cr})
@@ -38,10 +38,11 @@ feature_reference <-
            by.sound.file = FALSE,
            units = c("ms", "kHz"),
            digits = 2) {
-    if (is(reference, "extended_selection_table"))
+    if (is(reference, "extended_selection_table")) {
       stop2(
         "The function is not defined for class 'extended_selection_table'. Use 'selection_table' or 'data.frame' instea."
       )
+    }
 
     internal_feature_reference <-
       function(reference,
@@ -57,15 +58,15 @@ feature_reference <-
           mean(reference$duration, na.rm = TRUE)
         output$max.sel.duration <- max(reference$duration, na.rm = TRUE)
         suppressWarnings(output$min.gap.duration <-
-                           min(reference$gaps, na.rm = TRUE))
+          min(reference$gaps, na.rm = TRUE))
         suppressWarnings(output$mean.gap.duration <-
-                           mean(reference$gaps, na.rm = TRUE))
+          mean(reference$gaps, na.rm = TRUE))
         suppressWarnings(output$max.gap.duration <-
-                           max(reference$gaps, na.rm = TRUE))
+          max(reference$gaps, na.rm = TRUE))
 
-        if (total.annotations)
-          output$annotations <- nrow(reference) else
-        {
+        if (total.annotations) {
+          output$annotations <- nrow(reference)
+        } else {
           output$min.annotations <- min(count_annotations)
           output$mean.annotations <- mean(count_annotations)
           output$max.annotations <- max(count_annotations)
@@ -73,7 +74,7 @@ feature_reference <-
 
         # frequency range descriptors
         if (!is.null(reference$bottom.freq) &
-            !is.null(reference$top.freq)) {
+          !is.null(reference$top.freq)) {
           output$min.bottom.freq <- min(reference$bottom.freq, na.rm = TRUE)
           output$mean.bottom.freq <-
             mean(reference$bottom.freq, na.rm = TRUE)
@@ -86,11 +87,14 @@ feature_reference <-
 
         if (!is.null(path)) {
           durs <-
[…]

[…]
