+++ R/label_detection.R
@@ -68,6 +68,7 @@ label_detection <-
            cores = 1,
            pb = TRUE,
            min.overlap = 0.5) {
+
     if (is_extended_selection_table(reference)) {
       stop2("This function cannot take extended selection tables ('reference' argument)")
     }
@@ -106,7 +107,7 @@ label_detection <-
     }
 
     # no duplicated selection labels
-    if (anyDuplicated(paste(detection$sound.files, detection$selec)) > 0) {
+    if (anyDuplicated(paste(detection$sound.files, detection$selec, detection$template)) > 0) {
       stop2("Duplicated 'selec' labels within at least one sound file in 'detection'")
     }
 
@@ -156,6 +157,9 @@ label_detection <-
         cl = cl,
         X = unique(detection$sound.files),
         FUN = function(z) {
+
+          print(z)
+          #z <- "ch1MPI2020-01-23_10-30-35_0000106.wav"
           # get subset from detection for that sound file
           sub_detec <-
             as.data.frame(detection[detection$sound.files == z, ])
@@ -167,108 +171,112 @@ label_detection <-
               as.data.frame(reference[reference$sound.files == z, ])
 
             ovlp_events <- overlapping_detections(reference = sub_ref, detection = sub_detec)
+
             overlap_iou <- pairs_iou(df = ovlp_events, detection = sub_detec, reference = sub_ref)
 
-            # filter detections below minimum overlap
-            overlap_iou <- overlap_iou[overlap_iou$IoU > min.overlap, ]
+            if (nrow(overlap_iou) > 0){
 
-            sub_detec$detection.class <- vapply(paste(sub_detec$sound.files, sub_detec$selec, sep = "-"), function(y) {
-              # if no reference overlap is a false positive
-              if (sum(overlap_iou$detection.id == y) == 0) detection.class <- "false.positive"
+              # filter detections below minimum overlap
+              overlap_iou <- overlap_iou[overlap_iou$IoU > min.overlap, ]
 
-              # if only one overlapping reference is a true positive
-              if (sum(overlap_iou$detection.id == y) == 1) detection.class <- "true.positive"
+              sub_detec$detection.class <- vapply(paste(sub_detec$sound.files, sub_detec$selec, sep = "-"), function(y) {
+                # if no reference overlap is a false positive
+                if (sum(overlap_iou$detection.id == y) == 0) detection.class <- "false.positive"
 
-              # if more than one overlapping reference is a merged true positive
-              if (sum(overlap_iou$detection.id == y) > 1) detection.class <- "true.positive (merged)"
+                # if only one overlapping reference is a true positive
+                if (sum(overlap_iou$detection.id == y) == 1) detection.class <- "true.positive"
 
-              # if 1 reference overlaps and that reference overlaps with other detections
-              if (sum(overlap_iou$detection.id == y) == 1 & any(overlap_iou$reference.id[overlap_iou$detection.id != y] %in% overlap_iou$reference.id[overlap_iou$detection.id == y])) detection.class <- "true.positive (split)"
+                # if more than one overlapping reference is a merged true positive
+                if (sum(overlap_iou$detection.id == y) > 1) detection.class <- "true.positive (merged)"
 
-              if (sum(overlap_iou$detection.id == y) > 1 & any(overlap_iou$reference.id[overlap_iou$detection.id != y] %in% overlap_iou$reference.id[overlap_iou$detection.id == y])) detection.class <- "true.positive (split/merged)"
+                # if 1 reference overlaps and that reference overlaps with other detections
[…]
