+++ vignettes/ohun.Rmd
@@ -27,7 +27,7 @@ params:
 ---
 
 
-&nbsp; 
+ 
 
 <img src="ohun_sticker.png" alt="ohun sticker" align="right" width = "25%" height="25%"/>
 
@@ -35,9 +35,9 @@ params:
 
 
 
-&nbsp; 
+ 
 
-&nbsp; 
+ 
 
 <div class="alert alert-info">
 
@@ -46,7 +46,7 @@ params:
 - The use of reference annotations for detection optimization and diagnostic 
 - The use of signal detection theory indices to evaluate detection performance 
 
-&nbsp;
+
 
 <font size = "4">The package offers functions for: </font>
 
@@ -103,7 +103,7 @@ opts_chunk$set(
 )
 
 ```
-&nbsp;
+
 
 ---
 
@@ -116,11 +116,11 @@ Finding the position of sound events in a sound file is a challenging task. [ohu
 <figcaption><i>Diagram depicting how target sound event features can be used to tell the most adequate sound event detection approach. Steps in which 'ohun' can be helpful are shown in color. (SNR = signal-to-noise ratio) </i></figcaption>
 </figure> 
 
-&nbsp; 
+ 
 
 Still, a detection run using other software can be optimized with the tools provided in [ohun](https://github.com/maRce10/ohun). 
 
-&nbsp; 
+ 
 
 #  Signal detection theory applied to bioacoustics  
 
@@ -148,14 +148,14 @@ data("lbh1", "lbh2", "lbh_reference")
 
 lbh_reference
 ```
-&nbsp;
+
 
 This is a 'selection table', an object class provided by the package [warbleR](https://CRAN.R-project.org/package=warbleR) (see [`selection_table()`](https://marce10.github.io/warbleR/reference/selection_table.html) for details). Selection tables are basically data frames in which the contained information has been double-checked (using warbleR's [`check_sels()`](https://marce10.github.io/warbleR/reference/check_sels.html)). But they behave pretty much as data frames and can be easily converted to data frames:
 ```{r}
 # convert to data frame
 as.data.frame(lbh_reference)
 ```
-&nbsp;
+
 
 All [ohun](https://github.com/maRce10/ohun) functions that work with this kind of data can take both selection tables and data frames. Spectrograms with highlighted sound events from a selection table can be plotted with the function `label_spectro()` (this function only plots one wave object at the time):
 ```{r, eval = TRUE}
@@ -182,7 +182,7 @@ lbh1_reference <-
 diagnose_detection(reference = lbh1_reference, detection = lbh1_reference)[, c(1:3, 7:9)]
 
 ```
-&nbsp;
+
 
 We will work mostly with a single sound file for convenience but the functions can work on several sound files at the time. The files should be found in a single working directory. Although the above example is a bit silly, it shows the basic diagnostic indices, which include basic detection theory indices ('true.positives', 'false.positives', 'false.negatives', 'recall' and 'precision') mentioned above. We can play around with the reference table to see how these indices can be used to spot imperfect detection routines (and hopefully improve them!). For instance, we can remove some sound events to see how this is reflected in the diagnostics. Getting rid of some rows in 'detection', simulating a detection with some false negatives, will affect the recall but not the precision:
 ```{r}
@@ -205,7 +205,7 @@ label_spectro(
 diagnose_detection(reference = lbh1_reference, detection = lbh1_detection)[, c(1:3, 7:9)]
 
 ```
-&nbsp;
+
 
 Having some additional sound events not in reference will do the opposite, reducing precision but not recall. We can do this simply by switching the tables:
 ```{r, }
@@ -225,7 +225,7 @@ label_spectro(
[…]
