### 2023-05-31 161c808
rm commented output format
