### 2023-06-01 c5d5021
use expect_s3_class
