### 2023-06-01 73a3816
more expect s3 class
