### 2023-06-01 0dbcec3
use new f.score colname
