+++ tests/testthat/test-optimize_template_detector.R
@@ -23,7 +23,7 @@ test_that("1 template", {
 
   expect_equal(nrow(otp), 3)
 
-  expect_true(all(otp$f1.score > 0.68))
+  expect_true(all(otp$f.score > 0.68))
 })
 
 
@@ -62,5 +62,5 @@ test_that("2 templates", {
 
   expect_equal(nrow(otp), 6)
 
-  expect_true(all(otp$f1.score > 0.64))
+  expect_true(all(otp$f.score > 0.64))
 })
