### 2023-06-01 e0b5912
adjust spacing between text and chunks
