+++ vignettes/ohun.Rmd
@@ -144,7 +144,8 @@ Several additional indices derived from these indices are used to evaluate the p
 A perfect detection will have no false positives or false negatives, which will result in both recall and precision equal to 1. However, perfect detection cannot always be reached and some compromise between detecting all target signals plus some noise (recall = 1 & precision < 1) and detecting only target signals but not all of them (recall < 1 & precision = 1) is warranted. The right balance between these two extremes will be given by the relative costs of missing signals and mistaking noise for signals. Hence, these indices provide an useful framework for diagnosing and optimizing the performance of a detection routine. 
 
 The package [ohun](https://github.com/maRce10/ohun) provides a set of tools to evaluate the performance of an sound event detection based on the indices described above. To accomplish this, the result of a detection routine is compared against a reference table containing the time position of all target sound events in the sound files. The package comes with an example reference table containing annotations of long-billed hermit hummingbird songs from two sound files (also supplied as example data: 'lbh1' and 'lbh2'), which can be used to illustrate detection performance evaluation. The example data can be explored as follows:  
-```{r, eval = TRUE, }
+
+```{r, eval = TRUE}
 # load example data
 data("lbh1", "lbh2", "lbh_reference")
 
@@ -153,6 +154,7 @@ lbh_reference
 
 
 This is a 'selection table', an object class provided by the package [warbleR](https://CRAN.R-project.org/package=warbleR) (see [`selection_table()`](https://marce10.github.io/warbleR/reference/selection_table.html) for details). Selection tables are basically data frames in which the contained information has been double-checked (using warbleR's [`check_sels()`](https://marce10.github.io/warbleR/reference/check_sels.html)). But they behave pretty much as data frames and can be easily converted to data frames:
+
 ```{r}
 # convert to data frame
 as.data.frame(lbh_reference)
@@ -160,6 +162,7 @@ as.data.frame(lbh_reference)
 
 
 All [ohun](https://github.com/maRce10/ohun) functions that work with this kind of data can take both selection tables and data frames. Spectrograms with highlighted sound events from a selection table can be plotted with the function `label_spectro()` (this function only plots one wave object at the time):
+
 ```{r, eval = TRUE}
 # save sound file
 writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
@@ -175,8 +178,8 @@ label_spectro(wave = lbh2, reference = lbh_reference[lbh_reference$sound.files =
 ```
 
 The function `diagnose_detection()` evaluates the performance of a detection routine by comparing it to a reference table. For instance, a perfect detection is given by comparing `lbh_reference` to itself:
-```{r}
 
+```{r}
 lbh1_reference <-
   lbh_reference[lbh_reference$sound.files == "lbh1.wav",]
 
@@ -187,9 +190,8 @@ diagnose_detection(reference = lbh1_reference, detection = lbh1_reference)[, c(1
 
 
[…]
