### 2023-06-01 60ce950
show all warnings and messages
