### 2023-06-01 #568 @robitalec
-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

**A statement of need**

Please consider adding a section to the README that 1) clarifies how `ohun` fits into a broader workflow of working with sound event and 2) if any other packages or approaches are available that share similar functionality to `ohun`.

For 1), this could be adding few sentences describing what typical steps (and related packages) would come before and after using `ohun`. You could also add the workflow from the vignette to the README in this section. Consider a new user looking at the [Reference](https://marce10.github.io/ohun/reference/index.html) page on the website or the list of available functions in `ohun`... how do they fit together?

<summary>Click to expand</summary>

```r
library(ohun)
#> Loading required package: tuneR
#> Loading required package: warbleR
#> Loading required package: seewave
#> Loading required package: NatureSounds
#> Loading required package: knitr
lsf.str('package:ohun')
#> consensus_detection : function (detection, by = "overlap", filter = "max", cores = 1, pb = TRUE)
#> diagnose_detection : function (reference, detection, by.sound.file = FALSE, time.diagnostics = FALSE,
#>     cores = 1, pb = TRUE, path = NULL, by = NULL, macro.average = FALSE,
#>     min.overlap = 0.5)
#> energy_detector : function (files = NULL, envelopes = NULL, path = ".", hop.size = 11.6,
#>     wl = NULL, thinning = 1, bp = NULL, smooth = 5, threshold = 5, peak.amplitude = 0,
#>     hold.time = 0, min.duration = 0, max.duration = Inf, cores = 1, pb = TRUE)
#> feature_acoustic_data : function (path = ".", digits = 2)
#> feature_reference : function (reference, path = NULL, by.sound.file = FALSE, units = c("ms",
#>     "kHz"), digits = 2)
#> filter_detection : function (detection, by = "overlap", filter = "max", cores = 1, pb = TRUE)
#> get_envelopes : function (path = ".", files = NULL, bp = NULL, hop.size = 11.6, wl = NULL,
#>     cores = 1, thinning = 1, pb = TRUE, smooth = 5, normalize = TRUE)
#> get_templates : function (reference, acoustic.space = NULL, path = ".", n.sub.spaces = 1,
#>     plot = TRUE, color = "#21908C4D", ...)
#> label_detection : function (reference, detection, cores = 1, pb = TRUE, min.overlap = 0.5)
#> label_spectro : function (wave, reference = NULL, detection = NULL, envelope = FALSE, threshold = NULL,
#>     smooth = 5, collevels = seq(-100, 0, 5), palette = viridis::viridis,
#>     template.correlation = NULL, line.x.position = 2, hop.size = NULL,
#>     ...)
#> merge_overlaps : function (X, pb = TRUE, cores = 1)
#> optimize_energy_detector : function (reference, files = NULL, threshold = 5, peak.amplitude = 0, hop.size = 11.6,
#>     wl = NULL, smooth = 5, hold.time = 0, min.duration = NULL, max.duration = NULL,
#>     thinning = 1, cores = 1, pb = TRUE, by.sound.file = FALSE, bp = NULL,
#>     path = ".", previous.output = NULL, envelopes = NULL, macro.average = FALSE,
#>     min.overlap = 0.5)
#> optimize_template_detector : function (template.correlations, reference, threshold, cores = 1, pb = TRUE,
#>     by.sound.file = FALSE, previous.output = NULL, macro.average = FALSE,
#>     min.overlap = 0.5)
#> split_acoustic_data : function (path = ".", sgmt.dur = 10, sgmts = NULL, files = NULL, cores = 1,
#>     pb = TRUE, only.sels = FALSE, X = NULL)
#> summarize_diagnostic : function (diagnostic, time.diagnostics = FALSE, macro.average = FALSE)
#> template_correlator : function (templates, files = NULL, hop.size = 11.6, wl = NULL, ovlp = 0,
#>     wn = "hanning", cor.method = "pearson", cores = 1, path = ".", pb = TRUE,
#>     type = "fourier", fbtype = "mel", ...)
#> template_detector : function (template.correlations, cores = 1, threshold, pb = TRUE, verbose = TRUE)
```

For 2), as @sammlapp mentions, it's really helpful for a user to be aware of what the options are and how they differ. The README for the `gt` package has a good example of this here: https://github.com/rstudio/gt#how-gt-fits-in-with-other-packages-that-generate-display-tables

**Installation instructions**

The dependency `seewave` brings system dependencies that might not be immediately obvious to a new user of `ohun`. It might be helpful to include a link to `seewave`'s installing and troubleshooting system requirements page, for example: https://rug.mnhn.fr/seewave/inst.html

**Vignette(s)**

It might be worth splitting this vignette into multiple vignettes. At its current length, I find it a bit hard to digest all the moving pieces and options. I personally find it easier to learn and use R packages if there are many, shorter vignettes that describe the different major functions or steps offered. You could have maybe four vignettes: introduction, energy based detection, template based detection, and diagnosing/improving. The introductory could expand on how `ohun` integrates with other R packages in this field, and provide the background given in sections "Automatic sound event detection" and "Signal detection theory applied to bioacoustics". Then the following vignettes using the sections already written from the current vignette. This might just be my opinion - so if you feel like it is better for `ohun` users to have everything in one place, let me know.

Throughout the vignette, the warnings and messages were not shown using the chunk options warnings=FALSE and messages=FALSE. I would recommend leaving all messages and warnings in the rendered vignette, because a user may be surprised and confused to see them appear locally, when they are not shown in the vignette.

**Examples**

- Found errors locally using `devtools::run_examples()`

If recommendation to move `tuneR` to , the following issues can be resolved

In `feature_acoustic_data.R`:

- `could not find function "writeWave"` fixed by adding `library(tuneR)`

**Community guidelines**

Please consider also including the link to the repository (https://github.com/maRce10/ohun) in the URL field in the DESCRIPTION as it should have the most consistently up to date information for the package, for example in the case that the website is not updated.

**Packaging guidelines**

Looks good: package name, snake_case, object_verb function naming, citation, and code metadata. Consider using this following options for [keeping the code metadata up to date](https://github.com/ropensci/codemetar#keep-codemetajson-up-to-date) and for [keeping the CITATION file up to date](https://docs.ropensci.org/cffr/reference/cff_gha_update.html).

Code style improved with a [recent commit](https://github.com/maRce10/ohun/commit/6a32910cd5a794e274dc96ca44a1ca486434e075)
using the `styler` package. Before that, logical statements with `if` `else` were not easy to read.

It is not clear whether or not functions in `ohun` are pipeable (see the second point [here](https://devguide.ropensci.org/building.html#function-and-argument-naming)). If you think that is a relevant option, consider including an example in the vignette that shows this.

As mentioned above, the README could use more information about how `ohun` fits into the broader ecosystem, comparing `ohun` to other packages that provide related functionality and details on the system requirements for `seewave`. The README is also missing a brief demonstration of functionality (see the sixth point [here](https://devguide.ropensci.org/building.html#readme)). The long form examples in the vignettes are useful for a user that is already intending to use `ohun` but shorter examples in the README help new users understand what `ohun` is offering.

Documentation looks good with all functions having accompanying examples, detailed return values and references included in the vignette. There is good use of `@seealso` to link users to related functions in `ohun`. Only one URL needed editing (according to `urlchecker`), see below.

You could consider using `vdiffr` for testing plots. Currently the tests for functions that produce plots are just testing that they return NULL. This isn't a very useful test, and while testing plots is not straightforward or intuitive, you could see if `vdiffr` helps! More detailed comments on testing below.

Detailed comments on package dependencies below.

#### Additional comments

The following sections have further details from above, and sometimes an accompanying pull request. I found it easier to work through the review by looking at the source code locally, and making changes as I noticed things. I also felt like it would be easier to highlight specific lines of code by just editing them, instead of referring to eg. Lines 83-85 in some file. No pressure to accept any of these pull requests directly, and I hope that you find it a clear and useful approach.

##### Repository

The repository size is ~ 120 mb.

I recommend that you delete, at least, the following files:

- .Rproj.user/
- .Rhistory
- ..Rcheck/
- ..pdf
- .Rdata
- testing/
- docs/ (in this case since you are using the pkgdown workflow with the gh-pages branch)

Some of these files were presumably added before they were listed in the `.gitignore` file. I have deleted them in the following pull request:

https://github.com/maRce10/ohun/pull/12

##### README

Consider stating explicitly which package provides the parallel processing for `ohun` in the sentence starting "All functions allow the parallelization of tasks..."

The badge for `R-CMD-check` seems to point to an old workflow. The current workflow (`tic`) shows failing tests, whereas the `R-CMD-check` shows passing test. I would recommend removing this old badge if it is no longer relevant.

A couple minor edits to the README are in the following pull request:

https://github.com/maRce10/ohun/pull/13

##### NEWS

It might be helpful to users if the items in NEWS were accompanied with references to the relevant pull requests and issues (eg. #5). For example: https://github.com/tidyverse/ggplot2/blob/main/NEWS.md

##### Vignette

As mentioned above, consider splitting the single vignette into multiple vignettes:

- Get started vignette
    - Background and theory
    - How does ohun fit in
    - What other packages does ohun interface with
    - Overview of functions provided by ohun
- Energy based detection
- Template based detection
- FAQ / best practices / advice / tips

Minor proposed edits to the vignette are in the following pull request:

https://github.com/maRce10/ohun/pull/14

##### Contributing

The CONTRIBUTING.md document looks like it's based on the template from https://contributing.md/. Because of this, there were some places that links to sections didn't work, and it often still referred to the CONTRIBUTING.md version.

There isn't a code of conduct in the repository to link to, so I removed that dead link. But please consider including one (see the rOpenSci pages for more details: [Contributing Guide](https://devguide.ropensci.org/contributingguide.html), [Code of Conduct](https://ropensci.org/code-of-conduct/)).

Relevant pull request:

https://github.com/maRce10/ohun/pull/15

##### Documentation

I propose removing the "last modification date" from the R/ files since this information is redundant and more error prone than the information provided by Git.

##### DESCRIPTION

I suggest moving packages under "Depends" to "Imports". The case of adding packages to the Depends field does not seem like a universal yes/no clear cut decision. From my understanding and reading around this topic, I feel like for the `ohun` package, it is not necessary to include packages in the Depends field.

> Use Imports instead of Depends for packages providing functions from other packages.

See also https://github.com/leeper/Depends

For an example where it might be relevant to list a package in Depends:

> A more classic example of Depends is how the censored package depends on the parsnip and survival packages. Parsnip provides a unified interface for fitting models, and censored is an extension package for survival analysis. Censored is not useful without parsnip and survival, so it makes sense to list them in Depends.

In the case of `ohun`, some of the packages listed under Depends also list some packages under Depends:

- From `warbleR`
    - `tuneR`, `seewave` (>= 2.0.1), `NatureSounds`
- From `tuneR`
- From `seewave`
- From `NatureSounds`
    - `knitr`

The full dependency tree is here:

<summary>Click to expand</summary>

```
R: pak::pkg_deps_tree("ohun")
ohun 0.1.0 [new][bld][dl] (2.64 MB)
├─tuneR 1.4.3 [new][bld][cmp][dl] (427.79 kB)
│ └─signal 0.7-7 [new][bld][cmp][dl] (1.96 MB)
│   └─MASS 7.3-58.2 -> 7.3-59 [upd][bld][cmp][dl] (515.84 kB)
├─warbleR 1.1.28 [new][bld][cmp][dl] (4.49 MB)
│ ├─tuneR
│ ├─seewave 2.2.0 [new][bld][dl] (2.69 MB)
│ │ └─tuneR
│ ├─NatureSounds 1.0.4 [new][bld][dl] (4.40 MB)
│ │ ├─knitr 1.42 [new][bld][dl] (893.59 kB)
│ │ │ ├─evaluate 0.20 [new][bld][dl] (26.66 kB)
│ │ │ ├─highr 0.10 [new][bld][dl] (15.08 kB)
│ │ │ │ └─xfun 0.39 [new][bld][cmp][dl] (135.48 kB)
│ │ │ ├─yaml 2.3.7 [new][bld][cmp][dl] (94.33 kB)
│ │ │ └─xfun
│ │ └─tuneR
│ ├─dtw 1.23-1 [new][bld][cmp][dl] (746.72 kB)
│ │ └─proxy 0.4-27 [new][bld][cmp][dl] (74.62 kB)
│ ├─fftw 1.0-7 [new][bld][cmp][dl] (42.45 kB)
│ ├─monitoR 1.0.7 [new][bld][dl] (2.90 MB)
│ │ └─tuneR
│ ├─pbapply 1.7-0 [new][bld][dl] (23.81 kB)
│ ├─RCurl 1.98-1.12 [new][bld][cmp][dl] (731.48 kB)
│ │ └─bitops 1.0-7 [new][bld][cmp][dl] (10.81 kB)
│ ├─rjson 0.2.21 [new][bld][cmp][dl] (116.51 kB)
│ ├─Rcpp 1.0.10 [new][bld][cmp][dl] (2.94 MB)
│ ├─knitr
│ ├─crayon 1.5.2 [new][bld][dl] (40.57 kB)
│ └─bioacoustics 0.2.8 [new][bld][cmp][dl] (829.33 kB)
│   ├─htmltools 0.5.5 [new][bld][cmp][dl] (131.88 kB)
│   │ ├─digest 0.6.31 [new][bld][cmp][dl] (163.50 kB)
│   │ ├─base64enc 0.1-3 [new][bld][cmp][dl] (7.83 kB)
│   │ ├─rlang 1.1.1 [new][bld][cmp][dl] (762.53 kB)
│   │ ├─fastmap 1.1.1 [new][bld][cmp][dl] (46.41 kB)
│   │ └─ellipsis 0.3.2 [new][bld][cmp][dl] (8.07 kB)
│   │   └─rlang
│   ├─moments 0.14.1 [new][bld][dl] (7.64 kB)
│   ├─Rcpp
│   ├─stringr 1.5.0 [new][bld][dl] (175.70 kB)
│   │ ├─cli 3.6.1 [new][bld][cmp][dl] (567.29 kB)
│   │ ├─glue 1.6.2 [new][bld][cmp][dl] (106.51 kB)
│   │ ├─lifecycle 1.0.3 [new][bld][dl] (106.85 kB)
│   │ │ ├─cli
│   │ │ ├─glue
│   │ │ └─rlang
│   │ ├─magrittr 2.0.3 [new][bld][cmp][dl] (267.07 kB)
│   │ ├─rlang
│   │ ├─stringi 1.7.12 [new][bld][cmp][dl] (7.60 MB)
│   │ └─vctrs 0.6.2 [new][bld][cmp][dl] (964.63 kB)
│   │   ├─cli
│   │   ├─glue
│   │   ├─lifecycle
│   │   └─rlang
│   └─tuneR
├─pbapply
├─viridis 0.6.3 [new][bld][dl] (3.05 MB)
│ ├─viridisLite 0.4.2 [new][bld][dl] (1.27 MB)
│ ├─ggplot2 3.4.2 [new][bld][dl] (3.15 MB)
│ │ ├─cli
│ │ ├─glue
│ │ ├─gtable 0.3.3 [new][bld][dl] (130.06 kB)
│ │ │ ├─cli
│ │ │ ├─glue
│ │ │ ├─lifecycle
│ │ │ └─rlang
│ │ ├─isoband 0.2.7 [new][bld][cmp][dl] (1.59 MB)
│ │ ├─lifecycle
│ │ ├─MASS
│ │ ├─mgcv 1.8-42
│ │ │ ├─nlme 3.1-162
│ │ │ │ └─lattice 0.20-45 -> 0.21-8 [upd][bld][cmp][dl] (589.33 kB)
│ │ │ └─Matrix 1.5-3 -> 1.5-4 [upd][bld][cmp][dl] (2.18 MB)
│ │ │   └─lattice
│ │ ├─rlang
│ │ ├─scales 1.2.1 [new][bld][dl] (270.61 kB)
│ │ │ ├─farver 2.1.1 [new][bld][cmp][dl] (1.27 MB)
│ │ │ ├─labeling 0.4.2 [new][bld][dl] (10.16 kB)
│ │ │ ├─lifecycle
│ │ │ ├─munsell 0.5.0 [new][bld][dl] (182.65 kB)
│ │ │ │ └─colorspace 2.1-0 [new][bld][cmp][dl] (2.12 MB)
│ │ │ ├─R6 2.5.1 [new][bld][dl] (63.42 kB)
│ │ │ ├─RColorBrewer 1.1-3 [new][bld][dl] (11.64 kB)
│ │ │ ├─rlang
│ │ │ └─viridisLite
│ │ ├─tibble 3.2.1 [new][bld][cmp][dl] (565.98 kB)
│ │ │ ├─fansi 1.0.4 [new][bld][cmp][dl] (482.06 kB)
│ │ │ ├─lifecycle
│ │ │ ├─magrittr
│ │ │ ├─pillar 1.9.0 [new][bld][dl] (444.53 kB)
│ │ │ │ ├─cli
│ │ │ │ ├─fansi
│ │ │ │ ├─glue
│ │ │ │ ├─lifecycle
│ │ │ │ ├─rlang
│ │ │ │ ├─utf8 1.2.3 [new][bld][cmp][dl] (241.41 kB)
│ │ │ │ └─vctrs
│ │ │ ├─pkgconfig 2.0.3 [new][bld][dl] (6.08 kB)
│ │ │ ├─rlang
│ │ │ └─vctrs
│ │ ├─vctrs
│ │ └─withr 2.5.0 [new][bld][dl] (102.09 kB)
│ └─gridExtra 2.3 [new][bld][dl] (1.06 MB)
│   └─gtable
├─crayon
├─seewave
├─RCurl
├─fftw
├─knitr
├─rjson
├─rlang
├─sp 1.6-0 [new][bld][cmp][dl] (1.03 MB)
│ └─lattice
├─igraph 1.4.2 [new][bld][cmp][dl] (3.31 MB)
│ ├─magrittr
│ ├─Matrix
│ ├─pkgconfig
│ ├─rlang
│ └─cpp11 0.4.3 [new][bld][dl] (304.53 kB)
└─Sim.DiffProc 4.8 [new][bld][cmp][dl] (1.12 MB)
  ├─Deriv 4.1.3 [new][bld][dl] (37.21 kB)
  └─MASS
```

Minor suggested edits to the Description include:

- Removing non-standard fields including "Date/Publication",
"Repository", "Config/testthat/edition"
- Adding the link to the repository under URL
- Using line breaks to make the Description file easier

Relevant pull request:

https://github.com/maRce10/ohun/pull/16/

##### Tests

I suggested some edits to the test scripts to better take advantage of the variety and precision of available functions from `testthat`. Often, the tests were `expect_true` where you set up a logical statement, eg. `nrow(DT) == 3`. This works but when it fails, there is nothing informative about the error. Alternatively, if you set this test using `expect_length` then you would get a numeric difference between expectation and result. This might make it easier to identify why or how the test is failing.

I also suggest using a skip_on_windows function.

Writing tests can be really hard (and rewarding!) - I found these resources
helpful:

- https://r-pkgs.org/testing-design.html#what-to-test
- https://devguide.ropensci.org/building.html#testing
- https://mtlynch.io/good-developers-bad-tests/

Specifically in the first link "avoid testing simple code [...] test what you're not sure about, what's fragile, ...". This is a great opportunity for you to put your deep understanding about this kind of analysis to work, because you know when results don't look right, beyond the basic "is it a data.frame", etc. To this point, a lot of tests are just length checks - do you have any other expectations from the functions other than that they will return the expected length?

I've adjusted some of the tests where I noticed an alternative function that might be more informative. Here's the pull request:

https://github.com/maRce10/ohun/pull/17

Note: if packages are moved from Depends, it may require them to be explicitly loaded with `library()` in the tests where their functions are used.

##### R

I don't really understand why custom functions `stop2` and `warning2` are being used. They return the error or warning message, without the call being printed. Would this possibly make it harder for a new user to troubleshoot, if where the message originates from? `stop2` is used 105 times across 14 files and `warning2` is used 3 times across 3 files.

For example with `stop2`:

```r
R: label_detection(reference = matrix(lbh_reference[-1, ]), detection = lbh_reference)
Error: 'reference' is not of a class 'data.frame' or 'selection_table'
```

and with `stop`:

```r
R: label_detection(reference = matrix(lbh_reference[-1, ]), detection = lbh_reference)
Error in label_detection(reference = matrix(lbh_reference[-1, ]), detection = lbh_reference) :
  'reference' is not of a class 'data.frame' or 'selection_table'
```

Related, warnings are suppressed in R/feature_reference.R 3 times and in R/optimize_energy_detector.R 2 times.

Consider using `seq_len` instead of 1:nrow, 1:length, 1:ncol, ...

``` r
x <- data.frame(LETTERS[1:5])

1:nrow(x)
#> [1] 1 2 3 4 5

seq_len(nrow(x))
#> [1] 1 2 3 4 5

x <- data.frame()

1:nrow(x)
#> [1] 1 0

seq_len(nrow(x))
#> integer(0)
```

Related to dependencies described above under Description:

- internal functions from another package are used in 12 cases, though given all 12 are from `warbleR` and the maintainer of `ohun` is also the maintainer of `warbleR` this seems likely ok

[…]
