+++ R/get_templates.R
@@ -23,8 +23,8 @@
 #' @examples {
 #'   # Save example files into temporary working directory
 #'   data("lbh1", "lbh2", "lbh_reference")
-#'   writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
-#'   writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
+#'   tuneR::writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
+#'   tuneR::writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
 #'
 #'   # get a single mean template
 #'   template <- get_templates(reference = lbh_reference, path = tempdir())
