+++ R/template_correlator.R
@@ -33,12 +33,12 @@
 #'   data("lbh1", "lbh2", "lbh_reference")
 #'
 #'   # save sound files
-#'   writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
-#'   writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
+#'   tuneR::writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
+#'   tuneR::writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
 #'
 #'   # create template
 #'   templ <- lbh_reference[4, ]
-#'   templ2 <- selection_table(templ,
+#'   templ2 <- warbleR::selection_table(templ,
 #'     extended = TRUE, confirm.extended = FALSE,
 #'     path = tempdir()
 #'   )
@@ -73,7 +73,8 @@
 #'   )
 #'
 #'   # using an extended selection table
-#'   templ_est <- selection_table(templ, extended = TRUE, confirm.extended = FALSE, path = tempdir())
+#'   templ_est <- warbleR::selection_table(templ, extended = TRUE, confirm.extended = FALSE, 
+#'   path = tempdir())
 #'
 #'   tc_fr_est <- template_correlator(templates = templ_est, path = tempdir(), type = "fourier")
 #'
