+++ R/template_detector.R
@@ -20,8 +20,8 @@
 #'   data("lbh1", "lbh2", "lbh_reference")
 #'
 #'   # save sound files
-#'   writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
-#'   writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
+#'   tuneR::writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
+#'   tuneR::writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
 #'
 #'   # template for the first sound file in 'lbh_reference'
 #'   templ1 <- lbh_reference[1, ]
@@ -200,7 +200,7 @@ template_detector <-
     } else if (all(sel_table_df$sound.files %in% list.files(path = corr_call_path)) &
       all(!is.na(sel_table_df$start))) {
       sel_table_df <-
-        selection_table(
+        warbleR::selection_table(
           X = sel_table_df[!is.na(sel_table_df$start), ],
           path = corr_call_path,
           parallel = cores,
