+++ R/optimize_energy_detector.R
@@ -63,8 +63,8 @@
 #' @examples \donttest{
 #' # Save example files into temporary working directory
 #' data("lbh1", "lbh2", "lbh_reference")
-#' writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
-#' writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
+#' tuneR::writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
+#' tuneR::writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
 #'
 #' # using smoothing and minimum duration
 #' optimize_energy_detector(
@@ -152,7 +152,7 @@ optimize_energy_detector <-
     }
 
     # if reference is not a data frame
-    if (!any(is.data.frame(reference), is_selection_table(reference))) {
+    if (!any(is.data.frame(reference), warbleR::is_selection_table(reference))) {
       stop2("reference is not of a class 'data.frame' or 'selection_table'")
     }
 
