+++ R/feature_reference.R
@@ -15,8 +15,8 @@
 #' @examples {
 #'   # load data and save example files into temporary working directory
 #'   data("lbh1", "lbh2", "lbh_reference")
-#'   writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
-#'   writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
+#'   tuneR::writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
+#'   tuneR::writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
 #'
 #'   # summary across sound files
 #'   feature_reference(reference = lbh_reference, path = tempdir())
