### 2023-06-05 959bf88
Update template_based_detection.Rmd
