### 2023-06-08 b4facc7
Update template_based_detection.Rmd
