### 2023-06-09 98c47e6
remove source in vignettes
