+++ pkgdown.yml
@@ -2,8 +2,10 @@ pandoc: 2.19.2
 pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
-  ohun: ohun.html
-last_built: 2023-06-02T21:59Z
+  energy_based_detection: energy_based_detection.html
+  intro_to_ohun: intro_to_ohun.html
+  template_based_detection: template_based_detection.html
+last_built: 2023-06-09T00:21Z
 urls:
   reference: https://marce10.github.io/ohun/reference
   article: https://marce10.github.io/ohun/articles
