### 2023-06-09 7e550e3
Update energy_based_detection.Rmd
