+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   energy_based_detection: energy_based_detection.html
   intro_to_ohun: intro_to_ohun.html
   template_based_detection: template_based_detection.html
-last_built: 2023-06-09T00:37Z
+last_built: 2023-06-09T16:31Z
 urls:
   reference: https://marce10.github.io/ohun/reference
   article: https://marce10.github.io/ohun/articles
