+++ R/consensus_detection.R
@@ -1,6 +1,6 @@
 #' @title Remove ambiguous detections
 #'
-#' @description \code{consensus_detection} removes ambiguous detections (split and merged detections)
+#' @description \code{consensus_detection} removes ambiguous detections 
 #' @usage consensus_detection(detection, by = "overlap", filter = "max", cores = 1, pb = TRUE)
 #' @param detection Data frame or selection table (using the warbleR package's format, see \code{\link[warbleR]{selection_table}}) with the output of \code{\link{label_detection}} containing the start and end of the signals. Must contained at least the following columns: "sound.files", "selec", "start", "end", "detection.class" and "reference.row" (these last 2 are generated by \code{\link{label_detection}}). It must also contained the column indicated in the 'by' argument.
 #' @param by Character vector of length 1 indicating a column in 'detection' that will be used to filter delections. Must refer to a numeric column. Default is 'overlap', which is return by \code{\link{label_detection}}.
@@ -11,7 +11,7 @@
 #' @return A data frame or selection table (if 'detection' was also a selection table, warbleR package's format, see \code{\link[warbleR]{selection_table}}) as in 'X' but removing ambiguous detections (split and merged positives).
 #' @export
 #' @name consensus_detection
-#' @details This function removes ambiguous detections (split or merged detections, see \code{\link{diagnose_detection}}) keeping only the one that maximizes a criterium given by 'filter'. By default it keeps the detection with the highest overlap to the reference signal. It works on the output of \code{\link{label_detection}}.
+#' @details This function removes ambiguous detections keeping only the one that maximizes a criterium given by 'filter'. By default it keeps the detection with the highest overlap to the reference signal. It works on the output of \code{\link{label_detection}}. Useful when several detections match the same reference as in the case of template detection with multiple templates (see \code{\link{template_detector}}).
 #'
 #' @examples {
 #'   # load example data
@@ -57,7 +57,7 @@
 #' @references {
 #' #' Araya-Salas, M., Smith-Vidaurre, G., Chaverri, G., Brenes, J. C., Chirino, F., Elizondo-Calvo, J., & Rico-Guevara, A. 2022. ohun: an R package for diagnosing and optimizing automatic sound event detection. BioRxiv, 2022.12.13.520253. https://doi.org/10.1101/2022.12.13.520253
 #' }
-#' @seealso \code{\link{label_detection}}
+#' @seealso \code{\link{label_detection}}, \code{\link{template_detector}}
 #' @author Marcelo Araya-Salas (\email{marcelo.araya@@ucr.ac.cr}).
 # last modification on oct-31-2021 (MAS)
 
@@ -68,6 +68,7 @@ consensus_detection <-
            filter = "max",
            cores = 1,
            pb = TRUE) {
+    
     # save start time
     start_time <- proc.time()
 
@@ -104,16 +105,9 @@ consensus_detection <-
 
     # run loop over every detected signal in the reference
     filter_tp_list <-
-      warbleR:::pblapply_wrblr_int(X = unique(unlist(
-        sapply(true.positives$reference.row, function(x) {
-          unlist(strsplit(x, "-"))
-        }, USE.NAMES = FALSE)
-      )), cl = cl, pbar = pb, function(x) {
+      warbleR:::pblapply_wrblr_int(X = unique(true.positives$reference), cl = cl, pbar = pb, function(x) {
         # get those detection that overlapped with x
-        X <-
-          true.positives[vapply(true.positives$reference.row, function(y) {
-            any(unlist(strsplit(y, "-")) == x)
-          }, FUN.VALUE = logical(1)), ]
+        X <- true.positives[true.positives$reference == x, ]
 
         # order by 'by'
         X <- X[order(X[, by, drop = TRUE], decreasing = TRUE), ]
