### 2023-06-28 f857949
add vignettes to readme
