### 2023-07-06 9fdd163
added internal function check_arguments
