+++ R/optimize_template_detector.R
@@ -83,57 +83,24 @@ optimize_template_detector <-
            previous.output = NULL,
            macro.average = FALSE,
            min.overlap = 0.5) {
-    if (!is(template.correlations, "template_correlations")) {
-      stop2(
-        "'template.correlations' must be an object of class 'template_correlations' generated by template_detector()"
-      )
-    }
-
-    if (is_extended_selection_table(reference)) {
-      stop2("This function cannot take extended selection tables ('reference' argument)")
-    }
-
-    # if reference is not a data frame
-    if (!any(is.data.frame(reference), warbleR::is_selection_table(reference))) {
-      stop2("reference is not of a class 'data.frame' or 'selection_table'")
-    }
-
-    # check if all columns are found
-    if (any(!(
-      c("sound.files", "selec", "start", "end") %in% colnames(reference)
-    ))) {
-      stop2(paste(paste(
-        c("sound.files", "selec", "start", "end")[!(c(
-          "sound.files", "selec",
-          "start", "end"
-        ) %in% colnames(reference))],
-        collapse =
-          ", "
-      ), "column(s) not found in 'reference'"))
-    }
-
-    # if there are NAs in start or end stop
-    if (any(is.na(c(reference$end, reference$start)))) {
-      stop2("NAs found in start and/or end columns")
-    }
-
-    # if end or start are not numeric stop
-    if (any(
-      !is(reference$end, "numeric"),
-      !is(reference$start, "numeric")
-    )) {
-      stop2("'start' and 'end' must be numeric")
-    }
-
-    # if any start higher than end stop
-    if (any(reference$end - reference$start <= 0)) {
-      stop2(paste(
-        "Start is higher than or equal to end in",
-        length(which(reference$end - reference$start <= 0)),
-        "case(s)"
-      ))
-    }
-
+    
+    # check arguments
+    arguments <- as.list(base::match.call())
+    
+    # add objects to argument names
+    for(i in names(arguments)[-1])
+      arguments[[i]] <- get(i)
+    
+    # check each arguments
+    check_results <- check_arguments(fun = arguments[[1]], args = arguments)
+    
+    # report errors
+    checkmate::reportAssertions(check_results)
+    
+    # do not check arguments on internal ohun function here (energy_detector())
+    options(ohun_check_args = FALSE)
+    on.exit(options(ohun_check_args = TRUE))    
+    
     # check that all sound files in reference have and correlation vector
     if (!all(reference$sound.files %in% sapply(strsplit(names(
       template.correlations

[…]
