### 2023-07-06 1e34a85
added function plot_detection()
