+++ NAMESPACE
@@ -15,6 +15,7 @@ export(label_spectro)
 export(merge_overlaps)
 export(optimize_energy_detector)
 export(optimize_template_detector)
+export(plot_detection)
 export(split_acoustic_data)
 export(summarize_diagnostic)
 export(template_correlator)
@@ -48,6 +49,17 @@ importFrom(cli,make_ansi_style)
 importFrom(cli,num_ansi_colors)
 importFrom(cli,style_bold)
 importFrom(cli,style_italic)
+importFrom(ggplot2,aes)
+importFrom(ggplot2,facet_grid)
+importFrom(ggplot2,geom_point)
+importFrom(ggplot2,geom_polygon)
+importFrom(ggplot2,ggplot)
+importFrom(ggplot2,labs)
+importFrom(ggplot2,scale_color_manual)
+importFrom(ggplot2,scale_fill_manual)
+importFrom(ggplot2,scale_y_continuous)
+importFrom(ggplot2,theme_bw)
+importFrom(ggplot2,theme_classic)
 importFrom(igraph,E)
 importFrom(igraph,as_data_frame)
 importFrom(igraph,graph_from_incidence_matrix)
