### 2023-07-06 d4edd0d
add description of annotation format
