+++ R/energy_detector.R
@@ -4,7 +4,7 @@
 #' @usage energy_detector(files = NULL, envelopes = NULL, path = ".", hop.size = 11.6, wl = NULL,
 #' thinning = 1, bp = NULL, smooth = 5, threshold = 5, peak.amplitude = 0,
 #' hold.time = 0, min.duration = 0, max.duration = Inf, cores = 1, pb = TRUE)
-#' @param files Character vector indicating the sound files that will be analyzed. Optional. If 'files' and 'envelopes' are not supplied then the function will work on all supported format sound files in the working directory.
+#' @param files Character vector indicating the sound files that will be analyzed. Optional. If 'files' and 'envelopes' are not supplied then the function will work on all supported format sound files in the working directory. Supported file formats:'.wav', '.mp3', '.flac' and '.wac'.
 #' @param envelopes An object of class 'envelopes' (generated by \code{\link{get_envelopes}}) containing the amplitude envelopes of the sound files to be analyzed. If 'files' and 'envelopes' are not supplied then the function will work on all supported format sound files in the working directory.
 #' @param path Character string containing the directory path where the sound files are located.
 #' The current working directory is used as default.
