### 2023-07-06 7c47092
move internal functionto compute spectrogram cross-correlation outside template_correlator
