+++ R/internal_functions.R
@@ -511,3 +511,122 @@ check_arguments <- function(fun, args){
   invisible(NULL)
 }
 
+
+#### internal function to compute spectrogram cross-correlation
+# create function to calculate correlation between 2 spectrograms
+XC_FUN <- function(spc1, spc2, cm = cor.method) {
+  # define short and long envelope for sliding one (short) over the other (long)
+  if (ncol(spc1) > ncol(spc2)) {
+    lg.spc <- spc1
+    shrt.spc <- spc2
+  } else {
+    lg.spc <- spc2
+    shrt.spc <- spc1
+  }
+  
+  # get length of shortest minus 1 (1 if same length so it runs a single correlation)
+  shrt.lgth <- ncol(shrt.spc) - 1
+  
+  # steps for sliding one signal over the other
+  stps <- ncol(lg.spc) - ncol(shrt.spc)
+  
+  # set sequence of steps, if <= 1 then just 1 step
+  if (stps <= 1) {
+    stps <- 1
+  } else {
+    stps <- 1:stps
+  }
+  
+  # calculate correlations at each step
+  cors <- vapply(stps, function(x, cor.method = cm) {
+    warbleR::try_na(cor(
+      c(lg.spc[, x:(x + shrt.lgth)]),
+      c(shrt.spc),
+      method = cm,
+      use = "pairwise.complete.obs"
+    ))
+  }, FUN.VALUE = numeric(1))
+  
+  # make negative values 0
+  cors[cors < 0] <- 0
+  
+  return(cors)
+}
+
+
+##############################################################################################################
+
+#' print method for class \code{template_correlations}
+#'
+#' @param x Object of class \code{template_correlations}, generated by \code{\link{template_correlator}}.
+#' @param ...	 further arguments passed to or from other methods. Ignored when printing 'template_correlations' class objects.
+#' @keywords internal
+#'
+#' @export
+#' @returns Prints a summary of an object of class 'template_correlations'.
+
+
+print.template_correlations <- function(x, ...) {
+  message2(paste(
+    "Object of class",
+    cli::style_bold("'template_correlations' \n")
+  ))
+  
+  message2(
+    color = "silver", x =
+      paste(
+        "* The output of the following",
+        cli::style_italic("template_correlator()"),
+        "call: \n"
+      )
+  )
+  
+  cll <- paste0(deparse(x$call_info$call))
+  message2(color = "silver", x = cli::style_italic(gsub("    ", "", cll), "\n"))
+  
+  # get file and template names
+  files_templates <- sapply(names(x)[-length(x)], strsplit, "/")
+  templates <- unique(sapply(files_templates, "[[", 1))
+  files <- unique(sapply(files_templates, "[[", 2))
+  
+  message2(
+    color = "silver", x =
+      paste(
+        paste("* Contains", length(x) - 1, "correlation score vector(s) from"),
+        length(templates),
+        "template(s):\n",
+        paste(cli::style_italic(utils::head(templates), collapse = " ")),
+        if (length(templates) > 6) {
+          paste("... and", length(templates) - 6, "more")
+        } else {
+          ""
+        }
+      )
+  )
+  
+  message2(
+    color = "silver", x =
+      paste(
+        paste("\n... and"),
+        length(files),
+        "sound files(s):\n",
+        paste(cli::style_italic(utils::head(files), collapse = " ")),
+        if (length(files) > 6) {
+          paste("... and", length(files) - 6, "more")
+        } else {
+          ""
+        }
+      )
+  )
+  
+  # print ohun version
+  message2(
+    color = "silver", x =
+      paste0(
+        "\n * Created by ",
+        cli::style_bold("ohun "),
+        x$call_info$ohun.version
+      )
+  )
+}
+
