### 2023-07-06 4ca47bd
move detect_FUN outside energy_detector()
