+++ R/energy_detector.R
@@ -163,198 +163,6 @@ energy_detector <-
       files <- names(envelopes)[-length(envelopes)]
     }
     
-    # function for detecting sound events (i is the file name)
-    detect_FUN <-
-      function(file,
-               wl,
-               thres,
-               peak.amplitude,
-               min.duration,
-               max.duration,
-               path,
-               bp,
-               thinning,
-               smooth,
-               envlp) {
-        # get envelope if not supplied
-        if (is.null(envlp)) {
-          envlp <- env_ohun_int(
-            i = file,
-            path,
-            bp,
-            hop.size,
-            wl,
-            cores,
-            thinning,
-            pb,
-            smooth,
-            normalize = TRUE
-          )
-        } else {
-          envlp <- envelopes[[file]]
-        }
-        
-        # normalize to range
-        if (max(envlp$envelope) > 1) {
-          envlp$envelope <- envlp$envelope - min(envlp$envelope)
-          envlp$envelope <- envlp$envelope / max(envlp$envelope)
-        }
-        
-        # time interval between consecutive samples
-        hop.samples <- envlp$duration / (length(envlp$envelope) - 1)
-        
-        # get times at which threshold is crossed
-        cross_thresh <- unlist(lapply(2:length(envlp$envelope), function(x) {
-          # positive means going up
-          if (envlp$envelope[x] > thres & envlp$envelope[x - 1] <= thres) out <- hop.samples * (x - 1)
-          # negative means going down
-          if (envlp$envelope[x] <= thres & envlp$envelope[x - 1] > thres) out <- hop.samples * (x - 1) * -1
-          # anything else should be null to save memory
-          if (envlp$envelope[x] <= thres & envlp$envelope[x - 1] <= thres | envlp$envelope[x] > thres & envlp$envelope[x - 1] > thres) out <- NULL
-          
-          return(out)
-        }))
-        
-        ## FIX IF START OR END OF sound eventS IS NOT INCLUDED IN SOUND FILE
-        # get start and end of detections
-        # starts are the positive ones
-        starts <- cross_thresh[cross_thresh > 0]
-        # and ends the negative ones (should be converted to positive)
-        ends <- abs(cross_thresh[cross_thresh < 0])
-        
-        # if there is no end
-        if (length(starts) > 0 & length(ends) == 0) ends <- envlp$duration
-        
-        # if there is no start
-        if (length(ends) > 0 & length(starts) == 0) starts <- 0
-        
-        # if there are both starts and ends detected
-        if (length(starts) > 0 & length(ends) > 0) {
-          # if start is not lower in the first detection
-          if (starts[1] > ends[1]) starts <- c(0, starts)
-          
-          # if end is not higher in the last
-          if (starts[length(starts)] > ends[length(ends)]) ends <- c(ends, envlp$duration)
-          
-          # put time of detection in data frame
-          detections_df <-
-            data.frame(
-              sound.files = file,
-              duration = ends - starts,
-              selec = NA,
-              start = starts,
-              end = ends,
-              stringsAsFactors = FALSE
-            )
-          
-          # add row names
-          if (nrow(detections_df) > 0) {
-            detections_df$selec <- seq_len(nrow(detections_df))
-          }
-        } else { # return NAs
-          detections_df <-
-            data.frame(
-              sound.files = file,
-              duration = NA,
-              selec = NA,
-              start = NA,
-              end = NA,
-              stringsAsFactors = FALSE
-            )
-        }
-        
-        # TIME FILTERS
-        # if something was detected applied time filters
-        if (nrow(detections_df) > 0) {
-          ## HOLD TIME MERGING
-          # merge selections based on hold time
-          if (hold.time > 0 & nrow(detections_df) > 1) {
[…]
