+++ R/get_envelopes.R
@@ -18,7 +18,7 @@
 #' output (usually very large R objects / files). Default is \code{1} (no thinning). Higher sampling rates can afford higher size reduction (e.g. lower thinning values). Reduction is conducted by linear interpolation using \code{\link[stats]{approx}}. Note that thinning may decrease time precision and that the higher the thinning the less precise the time detection. It's generally not advised if no smoothing ('smooth' argument) is applied.
 #' @param pb Logical argument to control progress bar. Default is \code{TRUE}.
 #' @param smooth A numeric vector of length 1 to smooth the amplitude envelope
-#'   with a sum smooth function. It controls the time range (in ms) in which amplitude samples are smoothed (i.e. averaged with neighboring samples). Default is 5. 0 means no smoothing is applied. Note that smoothing is applied before thinning (see 'thinning' argument).
+#'   with a sum smooth function. It controls the time 'neighborhood' (in ms) in which amplitude samples are smoothed (i.e. averaged with neighboring samples). Default is 5. 0 means no smoothing is applied. Note that smoothing is applied before thinning (see 'thinning' argument). The function  \code{\link[warbleR]{envelope}} is used internally which is analogous to sum smoothing in code{\link[seewave]{env}}. This argument is used internally by \code{\link{get_envelopes}}.
 #' @param normalize Logical argument to control if envelopes are normalized to a 0-1 range.
 #' @return An object of class 'envelopes'.
 #' @export
@@ -235,99 +235,3 @@ print.envelopes <- function(x, ...) {
   # print ohun version
   message2(color = "silver", x = paste0("\n * Created by ", cli::style_bold("ohun "), x$call_info$ohun.version))
 }
-
-########################### internal function to get an envelope ###################
-
-env_ohun_int <-
-  function(i,
-           path,
-           bp,
-           hop.size,
-           wl,
-           cores,
-           thinning,
-           pb,
-           smooth,
-           normalize) {
-    # read wave object
-    wave_obj <- warbleR::read_sound_file(X = i, path = path)
-    
-    # adjust wl based on hope.size (convert to samples)
-    if (is.null(wl)) {
-      wl <- round(wave_obj@samp.rate * hop.size / 1000, 0)
-    }
-    
-    # make wl even if odd
-    if (!(wl %% 2) == 0) wl <- wl + 1
-    
-    # convert smooth to samples
-    smooth <- round(wave_obj@samp.rate * smooth / 1000, 0)
-    
-    # filter frequencies
-    if (!is.null(bp)) {
-      wave_obj <-
-        seewave::ffilter(
-          wave_obj,
-          f = wave_obj@samp.rate,
-          from = bp[1] * 1000,
-          to = bp[2] * 1000,
-          bandpass = TRUE,
-          wl = wl,
-          output = "Wave"
-        )
-    }
-    
-    # detect sound events based on amplitude (modified from seewave::timer function)
-    amp_vector <- wave_obj@left
-    
-    # original number of samples
-    n.samples <- length(amp_vector)
-    
-    # original duration
-    wave_dur <- seewave::duration(wave_obj)
-    
-    # extract envelope
-    envp <-
-      warbleR::envelope(
-        x = amp_vector,
-        ssmooth = smooth
-      )
-    
-    # flat edges (first and last 100 ms) if lower than lowest amp value
-    if (n.samples > wave_obj@samp.rate / 5) {
-      min.envp <- min(envp[(wave_obj@samp.rate / 10):(length(envp) - wave_obj@samp.rate / 5)])
-      
-      if (envp[1] < min.envp) envp[1:min(which(envp >= min.envp))] <- min.envp
-      
[…]
