### 2023-07-06 a00e687
messsage functions work on files in path by default
