### 2023-07-06 a16ff19
remove zero variance variable from PCA
