+++ R/get_templates.R
@@ -75,7 +75,11 @@ get_templates <-
       # remove columns with NAs
       spectral_parameters <-
         spectral_parameters[, !vapply(spectral_parameters, anyNA, FUN.VALUE = logical(1))]
-
+      
+      # remove columns with 0 variance
+      spectral_parameters <-
+        spectral_parameters[, vapply(spectral_parameters, function(x) if(is(x, "numeric")) var(x) else 1, FUN.VALUE = numeric(1)) > 0]
+      
       # get PCA
       pca <-
         stats::prcomp(spectral_parameters[, 2:ncol(spectral_parameters)], scale. = TRUE)
