### 2023-07-06 8c2675b
move internal function to internal_function.R
