+++ man/print.template_correlations.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/template_correlator.R
+% Please edit documentation in R/internal_functions.R
 \name{print.template_correlations}
 \alias{print.template_correlations}
 \title{print method for class \code{template_correlations}}

[…]
