+++ README.Rmd
@@ -40,7 +40,9 @@ The package offers functions for:
    - Optimize detection routines based on reference annotations
    - Energy-based detection
    - Template-based detection
-   
+  
+The implementation of detection diagnostics that can be applied to both built in detection methods and to those obtained from other software packages makes the package ohun an useful tool for conducting direct comparisons of the performance of different routines. The compatibility of 'ohun' with data formats already used by other sound analysis R packages (e.g. seewave,  warbleR) enables the integration of 'ohun' into more complex acoustic analysis workflows in a popular programming enviroment within the research community.
+ 
 All functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency. The package works on sound files in '.wav', '.mp3', '.flac' and '.wac' format.
 
 Install/load the package from CRAN as follows:
