### 2023-07-07 2dad204
Update NEWS.md
