### 2023-07-07 b237f58
seewave dependencies installation
