+++ tests/testthat/test-optimize_energy_detector.R
@@ -25,9 +25,10 @@ test_that("nothing to merge", {
 
   expect_equal(nrow(oed), 4)
 
-  expect_true(all(oed$recall == 1))
+  expect_true(all(oed$recall > 0.9))
 
-  expect_true(all(oed$precision > 0.22))
+  expect_true(all(oed$precision < 0.22))
+  expect_class(oed, "data.frame")
 })
 
 
@@ -56,7 +57,7 @@ test_that("including previous output in new call", {
     )
 
   expect_equal(nrow(oed), 6)
-
+  expect_class(oed, "data.frame")
   unlink(
     list.files(
       path = tempdir(),
