+++ R/feature_acoustic_data.R
@@ -23,7 +23,6 @@
 #' @references {
 #' Araya-Salas, M., Smith-Vidaurre, G., Chaverri, G., Brenes, J. C., Chirino, F., Elizondo-Calvo, J., & Rico-Guevara, A. 2022. ohun: an R package for diagnosing and optimizing automatic sound event detection. BioRxiv, 2022.12.13.520253. https://doi.org/10.1101/2022.12.13.520253
 #' }
-# last modification on jul-2022 (MAS)
 
 feature_acoustic_data <- function(path = ".", digits = 2) {
   
