+++ R/label_detection.R
@@ -63,7 +63,7 @@
 #' @references {
 #' Araya-Salas, M., Smith-Vidaurre, G., Chaverri, G., Brenes, J. C., Chirino, F., Elizondo-Calvo, J., & Rico-Guevara, A. 2022. ohun: an R package for diagnosing and optimizing automatic sound event detection. BioRxiv, 2022.12.13.520253. https://doi.org/10.1101/2022.12.13.520253
 #' }
-# last modification on jul-16-2021 (MAS)
+
 label_detection <-
   function(reference,
            detection,
