### 2023-07-07 #568 @maRce10
Apologies for the delay. Many thanks to both reviewers. I am really grateful for all your time and effort. I have made changes and responded to your comments/suggestions below, quoting the commit the change was made in.

## Response to @sammlapp

**comment:**  Vignette(s): demonstrating major functionality that runs successfully locally - I wasn't able to see or access the vignette in R studio with vignette()

 _**response:**  I have added a link to each of the 3 vignettes to README maRce10/ohun@f857949224ace5f75b1747b913ce98886676745e_

  **comment:**  I don't typically use R, so some things are not obvious to me. Where can I find the full API documentation? I opened the /docs/index.html to view the docs by rendering the HTML locally, but presumably they are online somewhere. I would recommend linking to the documentation near the top of the README.

_**response:**  I added a link to the package documentation to the top of README maRce10/ohun@1ffc441207235c1663e75a02f17a89c426e11290_

**comment:** The use of the word "reference" to mean "audio annotations" or "labels" is unfamiliar to me, I'm not sure if this word is typically used in the related warbleR package.

_**response:**  The use of the term reference to refer to the ground truth annotations seems to be common in the field (check for instance Mesaros, A., Heittola, T., Virtanen, T., & Plumbley, M. D. (2021). Sound event detection:
A tutorial. IEEE Signal Processing Magazine, 38(5), 67-83).
I prefer "reference" because you can also compare annotations from different software to check agreement. So it is not always manual annotations vs computer generated annotations._

**comment:** The package should give the user some expectation for under what circumstances the detection methods are likely to be effective. Although it mentions that high SNR is required for amplitude thresholding and stereotyped sounds are necessary for template matching, there are other requirements of the data for these approaches to be effective.

_**response:** I  added that time-frequency overlaps with other sounds can affect performance. I am not aware of any other main requirement. maRce10/ohun@d85c0e0d0b2b2310ad83f4c2ad687e3606b5fb95_

**comment:** to improve documentation and function/variable/argument names to better match the action that is actually taken by a function;

_**response:** The function filter_detection() has been renamed consensus_detection() maRce10/ohun@c1e108d2dd65b1c834ff767222ff5634e8f9742d_

**comment:** to consider a moderate refactor of code to increase modularity, reduce redundancy, and aim to follow the Single Responsibility Principle;

_**response:**  a modular approach was used for checking arguments. Note that all task that are shared by different functions are carried out by a unique function. For instance, diagnosing detection performance in optimize_template_detector() or optimize_energy_detector() is done by calling  diagnose_detection() and summarize_diagnostic(). diagnose_detection() also calls label_detection() which can also be called by users. get_templates() is called within template_detector(). Now all functions call check_arguments(). maRce10/ohun@c1e108d2dd65b1c834ff767222ff5634e8f9742d_

**comment:** increase code readability by breaking complex logical lines into separate lines with additional comments, plus comments describing the general purpose of large blocks of code.

_**response:**  I simplify the code in diagnose_detection()  and summarize_diagnostic() maRce10/ohun@69d81da42cc5b831779be1ebb3aaa2e794464877e_. Note that functions that deal with envelopes tend to avoid duplicating envelope objects to avoid using to much memory_

**comment:** Template-based cross correlation is implemented in [monitoR](https://cran.r-project.org/web/packages/monitoR/vignettes/monitoR_QuickStart.pdf), while energy-based detection is implemented in [bioacoustics](https://cran.r-project.org/web/packages/bioacoustics/vignettes/introduction.html) (the implementation seems more generic and featured than the one in ohun). I wonder if the proposed value of Ohun is mostly in its specific ability to interact with the formats of annotations produced by warbleR. If so, the top-level documentation and README could clarify this purpose.

_**response:** monitoR offers no way to evaluate detection performance. Ohun does, and can even be applied for optimizing detections._

**comment:** Specifically, when I started using the program without having any experience with warbleR, it was not clear to me how to create annotation tables or generally what the utility of most of the package functions are. Writing a description of the characteristics of data that this package expects as input would be helpful. For instance, "This package provides utilities for comparing detection and annotations of audio events described by frequency and time boxes. Creating annotations containing such frequency-time boxes [is possible in the warbleR package?]."

_**response:**  I added a paragraph similar to that proposed by the reviewer to README. maRce10/ohun@d4edd0df7249bffc59a9c7263ad7176a5c2d80cb_
Note that all argument descriptions clearly state the format of the annotations and they do not need to be created using warbleR. For instance the description of "reference" in all functions:_

`Data frame or 'selection.table' (following the warbleR package format) with the reference selections (start and end of the sound events) that will be used to evaluate the performance of the detection, represented by those selections in 'detection'. Must contained at least the following columns: "sound.files", "selec", "start" and "end".`

**comment:** In the language of rOpenSci software recommendations, this is a sign that the functions are not "modular". In other places it might be seen as a violation of the Single Responsibility Principle (SRP). For example, the label_specto function violates SRP because it doesn't just label a spectrogram: it computes a spectrogram, plots it, then overlays labels, and might additionally comput and plot an amplitude signal. As another example, I get_envelopes.R and energy_detector.R seem to both implement the process of getting an amplitude envelope, rather than using shared modular functionality specific to this purpose. Though it would take a significant refactoring effort, I believe creating a more modular code base that aims to achieve the SRP could improve the package and user experience overall.

_**response:**  The function label_spectro() was created for making the graphs in the vignette. This is stated in the description of the function in bold font ("Created for graphs included in the vignette, and probably only useful for that or for very short recordings"). So it is not call by any other function or calls any other ohun function. It just for the vignette. Not sure how to make this more evident. maRce10/ohun@f857949224ace5f75b1747b913ce98886676745e_

**comment:** In general, the code is difficult to read because of the formatting and smushing lots of logic into a single functional "line" (which is broken into many lines anyways due to formatting). Separating out logical steps into sequential lines, with comments where appropriate, would make the code easier to read/understand/maintain. However, it would take a major refactoring effort.

_**response:**  As mentioned above I simplify the code in diagnose_detection()  and summarize_diagnostic() which were the most cumbersome maRce10/ohun@69d81da42cc5b831779be1ebb3aaa2e794464877e_. Note that functions that deal with envelopes tend to avoid duplicating envelope objects to avoid using to much memory_

**comment:** Consider adding units to variable names eg hold_time_ms

_**response:**  I think that is a good suggestion although it is extremely uncommon in R packages so I would rather not include units in names to keep match what R users are expecting._

## split_acoustic_data.R:

**comment:**     documentation should list supported sound file types

_**response:**  I added supported formats to all function working on sound files maRce10/ohun@63f2be131f0b5e0a3be583892816657ba567cb66_

**comment:**     documentation should list required columns of table

_**response:**  The argument description already mentioned that: `X = 'selection_table' object or a data frame with columns
#' for sound file name (sound.files), selection number (selec), and start and end time of signal
#' (start and end).`_

**comment:**     documentation should say what will happen with incomplete audio clip at end of file

_**response:**  Not sure I follow.  I think there is only when thing that can happen: create the short clip (not really incomplete)_

**comment:**   comments denoting the purpose of larger sections of code are helpful, eg # INPUT VALIDATION
    or # Create the audio clips

_**response:**  I added more comments maRce10/ohun@05daadb5a7bf3ae60bec57c8c8c98d20d85d4884_

##summarize_diagnostic.R:

**comment:**  The code style makes it difficult to review the logic or infer the operation performed by much of this code. For example, lines 131-141 of summarize_diagnostic.R. In general I'm not reading this code base to check for logical errors, but I think that the code style will make it difficult for others (or the authors) to debug, extend, understand or modify the code. In this example, separating the logical operation into a few separate and sequential lines of code could make it easier to digest, and would provide opportunities to add comments for any individual steps whose purposes are not obvious. (For example in this code, a comment could explain why we would want a weighted mean of true positive durations.)

_**response:** as mention above I simplify the code to make it easier to understand (and debug)  maRce10/ohun@69d81da42cc5b831779be1ebb3aaa2e794464877e_

*comment:** Input validation is redundant across many functions, eg validating the cores argument. Rather than repeating code in several modules, use modular helper functions. That way, if you change something it is consistent across the codebase.
_**response:**  I added the internal  function check_arguments() which is now used by all other functions to validate arguments maRce10/ohun@9fdd16338cf
bda7cb096e9f3d6f737d6517fe714_

**comment:**  Comments should explain not just what's being done but also why (rather than re-stating the code in English). For example, instead of # set bottom and top freq if not supplied you could explain the bottom and top frequency are needed to make a spectrogram. if they're not supplied, we set them to 0 and Ns=sample_rate/2. Another example, in template_detector.R L177 "# if no detections" is not a helpful comment - the comment should explain what will happen and why.

_**response:** More comments were added. maRce10/ohun@05daadb5a7bf3ae60bec57c8c8c98d20d85d4884 and maRce10/ohun@4e3449d7d27eb697ec7e0be83af11c9607a5d45a_

**comment:** Since I'm not an R programmer, I might not understand conventions for writing R modules. However, I don't think it's a good idea to define your helper functions inside another function. For example L218 of template_correlator.R begins a definition of a function to create a spectrogram. Even if this is not part of the API (its a hidden or internal function), I think it should be defined outside of the template_correlator function (1) because it doesn't make sense that it would be re-defined each time template_correlator is called, and (2) because it makes reading the code for template_correlator unnecessarily long.

_**response:**  I moved the internal function out of template_correlator() maRce10/ohun@7c4709280fdb2f68a84d4463a086c4724272040c. Same thing for energy_detector() maRce10/ohun@4ca47bd993b95cecccfd09c490d3605137b1605d_

**comment:**  I'm not sure how namespaces work in R but it would be helpful for a code reader to know where functions are coming from. For instance, get_templates.R calls find_templates which is not explicitly imported or defined anywhere in the file. It apparently can be accessed because it exists in another .R file in the package, or because all functions from the package have been imported into the default namespace.

_**response:**  Internal functions are common in R. They can be called using namespace:::function. All internal functions in ohun are found in internal_functions.R_

## Notes on the Getting Started tutorial page

**comment:**  The "main features" description is jargon-y and could be written in broader language. For example, "The use of signal detection theory indices to evaluate detection performance" -> "Evaluate detection accuracy with precision, recall, and other metrics". As is, I feel like it obscures the functionality of the package. As another example, I wasn't sure what "Optimize detection routines based on reference annotations" meant - after reading further, I think it basically means "Adjust detection parameters to optimize accuracy on a labeled set of recordings"

_**response:**  I have included  "Adjust detection parameters to optimize accuracy " into the package description (maRce10/ohun@7617753447c4e23d97377a77286818cef84444e8). However, "Evaluate detection accuracy with precision, recall, and other metrics" assumes that users know what those metrics are. Most biologists working on bioacoustics are not familiar with those terms._

**comment:**  "Template-based detection" It's not true that template-based detection "doesn’t depend on the signal-to-noise ratio" - high background noise or low signal level will decrease the accuracy of this approach

_**response:**  I reworded this: "As it is less affected by signal-to-noise ratio it's more robust to higher levels of background noise" maRce10/ohun@3fd70a73dc78ab861f8d4cfd0f894386f9ccbf66_

**comment:** outputs (eg, tables) in the Getting Started tutorial are very hard to interpret in the current form, could they be rendered as human-readable tables?

_**response:**  I don't think humans cannot read them (I can). I agree they could more printed in a more friendly way but that would require adding more dependencies (e.g. kntr, kableExtra)_

**comment:** An example of how to analyze a large set of files, for instance all files in a folder-subfolder structure, would be helpful. I had to do some googling to figure out how to do "globbing" ie find all paths mathing a pattern.

_**response:**  I agree, but the amount of example data that can be included is very limited for R packages._

## Notes on API documentation and variable/function names

**comment:** Throughout the codebase, I recommend more informative variable and argument names: for example, in the function merge_overlaps, to a new user, pb is not interpreted as "progress bar"

_**response:**  I try to follow the same variable name use b other packages (including those creating progress bars)_

**comment:** Throughout the documentation: why are arguments that are numbers/strings called "numeric vectors of length 1" and "character vectors of length 1"? This makes me thing I should pass c(0) or [0] instead of 0, for instance, as an argument.

_**response:**  c(1) and 1 are equivalent and are both numeric vectors of length 1 (0 is not used as an index in R). Not sure how else I can call them._

### get_envelopes :

**comment:** smooth argument - should state the specific function/implementation of smoothing used

_**response:**  I added that is done using the function envelope from warbleR maRce10/ohun@6a2e11e1e8131411ac8c10de71ed090a57abc4cc_

**comment:**  paths and files arguments: using the current path as default is fragile and does not match the conventions used elsewhere in the package. I recommend making the argument(s) mandatory. It is also unclear that specifying a folder but not files means that the function should attempt to analyze any/all audio files in the folder (only a specific file extension? how does it know which files to analyze or not?). Is that what happens if files=NULL?_

_**response:** I agree that making path mandatory would be safer. However many R users like to set the working directory (using setwd()) before running analyses. That's why is common for functions to look into the current working directory by default. All 5 functions with the argument files use the same default value (NULL). I added that: "If not supplied the function will work on all sound files (in the supported format) in 'path'." maRce10/ohun@a00e687465eb57285f529544123a76718edc2629_

### energy_detector:

**comment:**  again, I don't think making files optional and defaulting to all audio in the working directory is a good idea.

_**response:**  Agree, but again it's friendly for most R users._

### label_spectro:

**comment:** can you provide an example of how to use this function with an audio file? If I want to run it on anything other than a simulated array of numbers or the two built-in audio files, I need some prior line of code to load the audio file,

_**response:**  As mentioned above, I made this function just to get a nice looking visualization of the data in the vignette. I don't think is very helpful for most real data. I have this info in bold in the description: "Created for graphs included in the vignette, and probably only useful for that or for very short recordings". However, people seem to get confused. I thought about  just having it inside the vignette without including it as a function in the package. However users always ask how those figures are made. So that's why I included it as a function._

**comment:**  bp argument (in several functions): how is the signal bandpassed? Should state specifically, for example, a 9th order Butterworth bandpass filter

_**response:**  I added that: "Bandpass is done using the function code{\link[seewave]{ffilter}}, which applies a short-term Fourier transformation to first create a spectrogram in which the target frequencies are filtered and then is back transform into a wave object using a reverse Fourier transformation." maRce10/ohun@59a984f0f5b14f6d06c219b376ad628577edbb6e_

### feature_reference:

**comment:** "Extract quantitative features of references" - from the function name and short description, I expected this to extract acoustic features of the annotated audio events themselves, rather than summary statistics of the annotations per se. Perhaps the description could be "Summarize temporal and frequency dimensions of annotations and gaps" or something similar. Ideally, the function's name would reflect its functionality, something like summarize_annotation_boxes.

_**response:**  I addopted the suggested description and added an alternative name for the function: summarize_reference() maRce10/ohun@e0f0c7591a4da491d6b29f386523d8a00ae21151_

**comment:** gap times: I tried using this summary to choose a hold.time argument for energy_detector, but the min, max, and mean are not very informative, because what matters is the typical gap time between close-together annotations. Maybe the median or 25th percentile, or some other statistic, would be more informative.

_**response:**  Yeah that is a tricky thing to get and depends a lot on duty cycle and signal duration. But I don't think there is a single metric that can get a good estimate. min and mean still
give you a sense of the range of gap values._

##  feature_acoustic_data

**comment:** Similarly, feature_acoustic_data sounds like it would measure some acoustic feature. Since the function returns information about the file types, I would call it something like summarize_audio_file_characteristics

_**response:** I added an alternative name for the function: summarize_acoustic_data() maRce10/ohun@6a54b30e4505fa4586ccc5220c3f796530f598b9_

## filter_detection

[…]
