### 2023-07-07 #568 @jhollist
Thanks for your responses to the reviews, @maRce10!

@sammlapp and @robitalec could you both take a look and let me know what you think?  You can use the reveiwer approval template to indicate your response: https://devdevguide.netlify.app/approval2template.html
