+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: ohun
 Title: Optimizing Acoustic Signal Detection
-Version: 0.1.1
+Version: 1.0.0
 Maintainer: Marcelo Araya-Salas <marcelo.araya@ucr.ac.cr>
 Description: Facilitates the automatic detection of acoustic signals, 
     providing functions to diagnose and optimize the performance of detection 
