### 2023-07-24 c392662
fix bug that happens inside baRulho
