### 2023-07-24 ff38d21
Update internal_functions.R
