### 2023-07-27 #568 @robitalec
Thanks @maRce10 for the detailed response, and for accepting those PRs. I am grateful that approach worked well for you too, I found it easier to look at logical chunks of changes as PRs personally.

## Reviewer Response
Following the same headers as above:

#### A statement of need

- It appears that while these changes have been added to the README.Rmd, they are not present in the README.md (and therefore not on the GitHub or pkgdown home page). Consider using a pre-commit hook to make sure these files stay synchronized (see `?usethis::use_readme_rmd()`)

#### Installation instructions

- same as above under "A statement of need", these changes are not reflected in the README, GitHub page, or website
### Packaging guidelines

Nice to see the {vdiffr} use.

Should {vdiffr} be added to the Suggests list instead of Depends like {testthat}? If so, you can remove the @importFrom call for `vdffir::expect_doppelganger`.

Relatedly though, the files under tests/testthat/_snaps are quite large:

- get-templates.svg 36.9 MB
- get-templates-env.svg 38.2 MB

Is there any way to reduce the size of these SVGs?

#### Additional comments

Some files I deleted in the accepted PR have resurfaced/are still present in the repository:
- .Rproj.user
- .Rhistory,
- docs/

#### stop2, warning2

Ok for removing the call. My initial concern was that it would be harder for a user to follow the traceback, but I see now that the .call argument doesn't impact that.  Since a user can still troubleshoot using the traceback, this should be fine.

``` r
stop2 <- function(...) {
    stop(..., call. = FALSE)
}
test <- function() {
    stop2('test error')
}
test()
#> Error: test error
traceback()
#> 3: stop(..., call. = FALSE) at #2
#> 2: stop2("test error") at #1
#> 1: test()
```

### New comments

#### Tests

Two directories are in the Rbuildignore (https://github.com/maRce10/ohun/blob/master/.Rbuildignore#L4-L5):

- testing (doesn't seem to be used)
- tests (added in this [commit](https://github.com/maRce10/ohun/commit/0eca94df3d5c95b6857b6a9be84016e44df99f42
))

This means that the tests aren't being run on CRAN

eg. https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-clang/ohun-00check.html

Compare this to eg. {dplyr}

https://www.r-project.org/nosvn/R.check/r-devel-linux-x86_64-debian-clang/dplyr-00check.html

and we can see the section "checking tests"...

Surprised CRAN didn't notice there were no tests in the submitted package, but maybe it isn't a requirement of CRAN packages I'm not sure. It would be good to have tests running on CRAN.

#### Vignettes

If we run `usethis::use_vignette("test")` we get the following changes:

```
✔ Setting active project to '/home/alecr/Documents/Local-git/ohun'
✔ Adding 'inst/doc' to '.gitignore'
✔ Adding '*.html', '*.R' to 'vignettes/.gitignore'
✔ Writing 'vignettes/test.Rmd'
• Modify 'vignettes/test.Rmd'
```

Most importantly, a file is written `vignettes/.gitignore` and to it two lines are added (ignoring rendered HTML and .R from the repo)

```
*.html
*.R
```

You could consider deleting files that correspond to this wildcard then adding this new `.gitignore` file.

Two options I can think of for reducing the size of your vignettes:

1. Adjust your output options

For example, the original output options for `intro_to_ohun.Rmd` (8 MB)

```
output:
  rmarkdown::html_document:
    css: extra.css
    theme: null
    self_contained: yes
    toc: true
    toc_depth: 3
    toc_float:
      collapsed: false
      smooth_scroll: true
```

8 MB -> 7.1 MB using theme: null and removing toc_float

```
output:
  rmarkdown::html_document:
    css: extra.css
    self_contained: yes
    toc: true
    toc_depth: 3
```

7.1 MB -> 3.4 MB dropping css (might be due to fonts)

```
output:
  rmarkdown::html_document:
    theme: null
    self_contained: yes
    toc: true
    toc_depth: 3
```

3.4 MB --> 2.5 MB switching from output type html_document to html_vignette

```
output:
  rmarkdown::html_vignette:
    self_contained: yes
    toc: true
    toc_depth: 3
```

You could keep these changes from CRAN, then apply the CSS independently to your site?

2. Or you could consider using articles instead of vignettes: https://r-pkgs.org/vignettes.html#sec-vignettes-article

---

Here is my author chunk from other packages.

```
    c(person(given = "Alec L.",
             family = "Robitaille",
             role = ,
             email = "robit.alec@gmail.com",
             comment = c(ORCID = "0000-0002-4706-1762"))
```

---

Let me know what you think of these additional comments and responses. Thank you!
