### 2023-08-08 #568 @jhollist
@maRce10 Can you take a look at @robitalec responses and revise/respond to each?

@sammlapp could you take a look at the revisions and let me know what you think? You can use the reviewer approval template to indicate your response: https://devdevguide.netlify.app/approval2template.html

Thanks!
