+++ R/diagnose_detection.R
@@ -97,33 +97,32 @@ diagnose_detection <-
            by = NULL,
            macro.average = FALSE,
            min.overlap = 0.5) {
-    
     # check arguments
-    if (options("ohun_check_args")$ohun_check_args){
-      
+    if (options("ohun_check_args")$ohun_check_args) {
       # check arguments
       arguments <- as.list(base::match.call())
-      
+
       # add objects to argument names
-      for(i in names(arguments)[-1])
+      for (i in names(arguments)[-1]) {
         arguments[[i]] <- get(i)
-      
+      }
+
       # check each arguments
       check_results <- check_arguments(fun = arguments[[1]], args = arguments)
-      
+
       # report errors
       checkmate::reportAssertions(check_results)
     }
-    
+
     # do not check arguments on internal ohun function here (label_detection())
     options(ohun_check_args = FALSE)
-    on.exit(options(ohun_check_args = TRUE))    
-    
+    on.exit(options(ohun_check_args = TRUE))
+
     # do it by
     # run the function for each subset split by "by"
     if (!is.null(by)) {
       split_det <- split(x = detection, f = detection[, by])
-      
+
       split_diagnostic <-
         warbleR:::pblapply_wrblr_int(
           X = seq_len(length(split_det)),
@@ -141,45 +140,45 @@ diagnose_detection <-
                 by.sound.file = by.sound.file,
                 macro.average = macro.average
               )
-            
+
             # add by label
             by_diag$by <- names(split_det)[x]
-            
+
             # order columns
             by_diag <- by_diag[, c(ncol(by_diag), 1:(ncol(by_diag) - 1))]
-            
+
             # rename by column
             names(by_diag)[1] <- by
-            
+
             return(by_diag)
           }
         )
-      
+
       performance_df <- do.call(rbind, split_diagnostic)
     } else {
       # make it a data frame if selection table
-      if ( warbleR::is_selection_table(detection)) {
+      if (warbleR::is_selection_table(detection)) {
         detection <- as.data.frame(detection)
       }
-      
+
       # make it a data frame if selection table
-      if ( warbleR::is_selection_table(reference)) {
+      if (warbleR::is_selection_table(reference)) {
         reference <- as.data.frame(reference)
       }
-      
+
       # remove rows with no info
       detection <- detection[!is.na(detection$start), ]
-      
+
       # message if more sound files in detection than in reference
       extra_detec_sf <-
         setdiff(detection$sound.files, reference$sound.files)
-      
+
       if (length(extra_detec_sf)) {
         on.exit(warning(
           "There is at least one additional sound file in 'detection' not found in 'reference'"
         ))
       }
-      
+
       if (nrow(detection) > 0) {
         # double checking happens inside label_detection()
         labeled_detection <-
@@ -190,11 +189,11 @@ diagnose_detection <-
             pb = pb,
             min.overlap = min.overlap
           )
-        
-        
+
+
         # add overlaps as attributes
         overlaps <- attributes(labeled_detection)$overlaps
-        
+
         # look at detections matching 1 reference selection at the time
         performance_list <-
           lapply(unique(labeled_detection$sound.files), function(z) {
@@ -205,10 +204,10 @@ diagnose_detection <-
             # get subset of detections for that sound file
             sub_ref <- reference[reference$sound.files == z, ]
             sub_ref$id <- paste(sub_ref$sound.files, sub_ref$selec, sep = "-")
-            
+
             # get subset of overlaps for that sound file
             sub_overlaps <- overlaps[overlaps$sound.files == z, ]
-            
+
             # put all performance indices in a data frame
             performance <- data.frame(
               sound.files = z,
@@ -227,10 +226,10 @@ diagnose_detection <-
                 "merged", sub_detec$detection.class
               )),
[…]
