+++ R/consensus_detection.R
@@ -1,6 +1,6 @@
 #' @title Remove ambiguous detections
 #'
-#' @description \code{consensus_detection} removes ambiguous detections 
+#' @description \code{consensus_detection} removes ambiguous detections
 #' @usage consensus_detection(detection, by = "overlap", filter = "max", cores = 1, pb = TRUE)
 #' @param detection Data frame or selection table (using the warbleR package's format, see \code{\link[warbleR]{selection_table}}) with the output of \code{\link{label_detection}} containing the start and end of the signals. Must contained at least the following columns: "sound.files", "selec", "start", "end" and "detection.class" (the last one is generated by \code{\link{label_detection}}). It must also contained the column indicated in the 'by' argument (which is 'overlap' by default).
 #' @param by Character vector of length 1 indicating a column in 'detection' that will be used to filter delections. Must refer to a numeric column. Default is 'overlap', which is return by \code{\link{label_detection}}.
@@ -67,93 +67,93 @@ consensus_detection <-
            filter = "max",
            cores = 1,
            pb = TRUE) {
-    
     # save start time
     start_time <- proc.time()
-    
+
     # check arguments
     arguments <- as.list(base::match.call())
-    
+
     # add objects to argument names
-    for(i in names(arguments)[-1])
+    for (i in names(arguments)[-1]) {
       arguments[[i]] <- get(i)
-    
+    }
+
     # check each arguments
     check_results <- check_arguments(fun = arguments[[1]], args = arguments)
-    
+
     # report errors
     checkmate::reportAssertions(check_results)
-    
+
     # check if by column is found
     if (!by %in% names(detection)) {
       stop2("'by' column not found")
     }
-    
+
     # add row id column to la
     detection$..row.id <- seq_len(nrow(detection))
-    
+
     # split in false and true positives
     false.positives <-
       as.data.frame(detection[grep("false.positive", detection$detection.class), ])
     true.positives <-
       as.data.frame(detection[grep("true.positive", detection$detection.class, FALSE), ])
-    
+
     # set clusters for windows OS
     if (Sys.info()[1] == "Windows" & cores > 1) {
       cl <- parallel::makeCluster(cores)
     } else {
       cl <- cores
     }
-    
+
     # run loop over every detected signal in the reference
     filter_tp_list <-
       warbleR:::pblapply_wrblr_int(X = unique(true.positives$reference), cl = cl, pbar = pb, function(x) {
         # get those detection that overlapped with x
         X <- true.positives[true.positives$reference == x, ]
-        
+
         # order by 'by'
         X <- X[order(X[, by, drop = TRUE], decreasing = TRUE), ]
-        
+
         # filter
         if (filter == "max") {
           X <- X[1, , drop = FALSE]
         } else {
           X <- X[nrow(X), , drop = FALSE]
         }
-        
+
         return(X)
       })
-    
+
     # put together in a data frame
     filter_tp_df <- do.call(rbind, filter_tp_list)
-    
+
     # add false positives
     filtered_detection <- rbind(false.positives, filter_tp_df)
-    
+
     # sort back
     filtered_detection <-
       filtered_detection[order(filtered_detection$..row.id), ]
-    
+
     # convert back to selection table
     if (warbleR::is_selection_table(detection)) {
       # keep only those rows in filtered_detections
       detection <-
         detection[detection$..row.id %in% filtered_detection$..row.id, ]
       detection <- detection[order(detection$..row.id), ]
-      
+
       # overwrite labeled_detections
       filtered_detection <- detection
-      
+
       # fix call
       attributes(filtered_detection)$call <- base::match.call()
-      
+
       # add elapsed time
[…]
