### 2023-08-31 50b7a55
add contributer and move vdiffr to "suggests"
