+++ DESCRIPTION
@@ -10,14 +10,19 @@ License: GPL (>= 2)
 Encoding: UTF-8
 URL: https://marce10.github.io/ohun/, https://github.com/maRce10/ohun
 BugReports: https://github.com/maRce10/ohun/issues/
-VignetteBuilder: knitr, rmarkdown
+VignetteBuilder: knitr
 RoxygenNote: 7.2.3
 Language: en-US
 Authors@R: 
     c(person("Marcelo", "Araya-Salas", 
       role = c("aut", "cre"), 
       email = "marcelo.araya@ucr.ac.cr", 
-      comment = c(ORCID = "0000-0003-3594-619X")))
+      comment = c(ORCID = "0000-0003-3594-619X")),
+      person(given = "Alec L.",
+      family = "Robitaille",
+      role = "ctb",
+      email = "robit.alec@gmail.com",
+      comment = c(ORCID = "0000-0002-4706-1762")))       
 Imports: tuneR, 
     warbleR (>= 1.1.28), 
     cli, 
@@ -30,11 +35,11 @@ Imports: tuneR,
     sf, 
     igraph,
     checkmate,
-    ggplot2,
-    vdiffr
+    ggplot2
 Depends: R (>= 3.2.1)
 Suggests: knitr, 
     rmarkdown, 
     testthat, 
     viridis, 
-    Sim.DiffProc
+    Sim.DiffProc,
+    vdiffr
