### 2023-08-31 8bc198a
add gitignore to vignette
