### 2023-08-31 #568 @maRce10
Hi @robitalec

Thanks again! I added you as a contributor and move vdiffr to "suggests" (maRce10/ohun@50b7a55b12a2d13ce2865ccb2e0f6f4b8afea8f2), render the readme (maRce10/ohun@7f530964cedea39a091ab476f03c7079fe9716c0) and added a vignette gitignore file (maRce10/ohun@8bc198ac25938829bb1e0e4976a48ed12037e52b).

Will change the vignette size before submitting to CRAN. Any suggestion on how to keep the 2 versions of the vignettes (CRAN and site) without having to change manually the yaml header?

cheers
