### 2023-09-08 #568 @robitalec
Thanks @maRce10!
Looks good, thanks for considering the suggestions.

Thanks for clarifying @jhollist that the vignettes do not need to adhere to CRAN size suggestions.

## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 17
