+++ DESCRIPTION
@@ -20,11 +20,16 @@ Authors@R:
       comment = c(ORCID = "0000-0003-3594-619X")),
       person(given = "Alec L.",
       family = "Robitaille",
-      role = "ctb",
+      role = "rev",
       email = "robit.alec@gmail.com",
-      comment = c(ORCID = "0000-0002-4706-1762")))       
+      comment = c(ORCID = "0000-0002-4706-1762")),
+      person(given = "Sam",
+      family = "Lapp",
+      role = "rev",
+      email = "sammlapp@gmail.com",
+      comment = c(ORCID = "0000-0003-1637-6822")))       
 Imports: tuneR, 
-    warbleR (>= 1.1.28), 
+    warbleR (>= 1.1.29), 
     cli, 
     methods, 
     stats, 
