+++ inst/doc/intro_to_ohun.R
@@ -1,192 +1,0 @@
-params <-
-list(EVAL = FALSE)
-
-## ---- eval = FALSE----------------------------------------------------------------------
-#  
-#  # From CRAN would be
-#  install.packages("ohun")
-#  
-#  #load package
-#  library(ohun)
-#  
-
-## ---- eval = FALSE----------------------------------------------------------------------
-#  
-#  # install package
-#  remotes::install_github("maRce10/ohun")
-#  
-#  #load packages
-#  library(ohun)
-#  library(tuneR)
-#  library(warbleR)
-
-## ----global options, echo = FALSE, message=FALSE, warning=FALSE-------------------------
-
-# library(checkmate)
-# library(ggplot2)
-# source("~/Dropbox/R_package_testing/ohun/R/plot_detection.R")
-# source("~/Dropbox/R_package_testing/ohun/R/internal_functions.R")
-# 
-# 
-# load("../data/lbh2.rda")
-# load("../data/lbh1.rda")
-# load("../data/lbh_reference.rda")
-
-#load packages
-library(ohun)
-library(tuneR)
-library(warbleR)
-library(ggplot2)
-
-data("lbh1", "lbh2", "lbh_reference")
-
-# for spectrograms
-par(mar = c(5, 4, 2, 2) + 0.1)
-
-stopifnot(require(knitr))
-options(width = 90)
-opts_chunk$set(
-  comment = NA,
-  # eval = if (isTRUE(exists("params"))) params$EVAL else FALSE,
-  dev = "jpeg",
-  dpi = 100,
-  fig.width=10,
-  out.width = "100%",
-  fig.align = "center"
-)
-
-
-## ---- eval = TRUE-----------------------------------------------------------------------
-# load example data
-data("lbh1", "lbh2", "lbh_reference")
-
-lbh_reference
-
-## ---------------------------------------------------------------------------------------
-# convert to data frame
-as.data.frame(lbh_reference)
-
-## ---- eval = TRUE, fig.asp=0.4----------------------------------------------------------
-# save sound file
-tuneR::writeWave(lbh1, file.path(tempdir(), "lbh1.wav"))
-
-# save sound file
-tuneR::writeWave(lbh2, file.path(tempdir(), "lbh2.wav"))
-
-# print spectrogram
-label_spectro(wave = lbh1, reference = lbh_reference[lbh_reference$sound.files == "lbh1.wav", ], hop.size = 10, ovlp = 50, flim = c(1, 10))
-
-# print spectrogram
-label_spectro(wave = lbh2, reference = lbh_reference[lbh_reference$sound.files == "lbh2.wav", ], hop.size = 10, ovlp = 50, flim = c(1, 10))
-
-## ---------------------------------------------------------------------------------------
-lbh1_reference <-
-  lbh_reference[lbh_reference$sound.files == "lbh1.wav",]
-
-# diagnose
-diagnose_detection(reference = lbh1_reference, detection = lbh1_reference)[, c(1:3, 7:9)]
-
-
-## ---- fig.asp=0.4-----------------------------------------------------------------------
-# create new table
-lbh1_detection <- lbh1_reference[3:9,]
-
-# print spectrogram
-label_spectro(
-  wave = lbh1,
-  reference = lbh1_reference,
-  detection = lbh1_detection,
-  hop.size = 10,
-  ovlp = 50,
-  flim = c(1, 10)
-)
-
-# diagnose
-diagnose_detection(reference = lbh1_reference, detection = lbh1_detection)[, c(1:3, 7:9)]
-
-
-## ---- fig.asp=0.4-----------------------------------------------------------------------
-# print spectrogram
[…]
